#15474 [SuperBr] [mod-sei-eouv] Melhorias no código
