+++ sei/web/modulos/cgu/mod-sei-eouv/rn/MdCguEouvAgendamentoRN.php
@@ -11,6 +11,8 @@ error_reporting(E_ALL); ini_set('display_errors', '1');
 
 require_once dirname(__FILE__) . '/../../../../SEI.php';
 
+//header('Content-Type: text/html; charset=UTF-8');
+
 class MdCguEouvAgendamentoRN extends InfraRN
 {
     protected $urlWebServiceEOuv;
@@ -33,16 +35,375 @@ class MdCguEouvAgendamentoRN extends InfraRN
     protected $dataInicialImportacaoManifestacoes;
     protected $identificacaoServico = 'CadastrarManifestacao';
     protected $siglaSistema = 'EOUV';
-
-
+    
     public function __construct()
     {
         parent::__construct();
+        //ini_set('memory_limit', '1024M');
     }
 
     protected function inicializarObjInfraIBanco()
     {
         return BancoSEI::getInstance();
+    }
+
+    /**
+     * Função para importar as manifestações e-Ouv do FalaBR
+     *
+     * Tipos: 1, 2, 3, 4, 5, 6 e 7
+     */
+    public function executarImportacaoManifestacaoEOuv()
+    {
+        // Log
+        LogSEI::getInstance()->gravar('Rotina de Importação de Manifestações do E-Ouv', InfraLog::$INFORMACAO);
+
+        $objEouvParametroDTO = new MdCguEouvParametroDTO();
+        $objEouvParametroDTO -> retTodos();
+
+        // Busca parâmetros do banco de dados
+        $objEouvParametroRN = new MdCguEouvParametroRN();
+        $arrObjEouvParametroDTO = $objEouvParametroRN->listarParametro($objEouvParametroDTO);
+
+        $numRegistros = count($arrObjEouvParametroDTO);
+
+        $this->preencheVariaveisEouv($numRegistros, $arrObjEouvParametroDTO);
+
+        /**
+         * Função para buscar o 'restante' do token sem o limite de 255 caracteres do SEI
+         */
+        $tokenPart2 = BancoSEI::getInstance()->consultarSql('select substring(de_valor_parametro, 256, 455) from md_eouv_parametros where id_parametro=10;')[0]['computed'];
+        $this->token = $this->token . $tokenPart2;
+
+        $objInfraParametro = new InfraParametro(BancoSEI::getInstance());
+        $idUsuarioSei = $objInfraParametro->getValor('ID_USUARIO_SEI');
+        $dataAtual = InfraData::getStrDataHoraAtual();
+
+        $isBolHabilitada = SessaoSEI::getInstance(false)->isBolHabilitada();
+        SessaoSEI::getInstance()->setBolHabilitada(false);
+
+        // Simula login inicial
+        self::simulaLogin($this->siglaSistema, $this->identificacaoServico, $this->idUnidadeOuvidoria);
+
+        try {
+
+            //Retorna dados da Última execução com Sucesso
+            $objUltimaExecucao = MdCguEouvAgendamentoINT::retornarUltimaExecucaoSucesso();
+
+            if ($objUltimaExecucao != null) {
+                $ultimaDataExecucao = $objUltimaExecucao->getDthDthPeriodoFinal();
+            } //Primeira execução ou nenhuma executada com sucesso
+            else {
+                $ultimaDataExecucao = $this->dataInicialImportacaoManifestacoes;
+            }
+
+            $semManifestacoesEncontradas = true;
+            $qtdManifestacoesNovas = 0;
+            $qtdManifestacoesAntigas = 0;
+            $objEouvRelatorioImportacaoDTO = $this->gravarLogImportacao($ultimaDataExecucao, $dataAtual);
+            $idRelatorioImportacao = $objEouvRelatorioImportacaoDTO->getNumIdRelatorioImportacao();
+            $objEouvRelatorioImportacaoRN = new MdCguEouvRelatorioImportacaoRN();
+            $SinSucessoExecucao = 'N';
+            $textoMensagemErroToken = '';
+
+            $retornoWs = $this->executarServicoConsultaManifestacoes($this->urlWebServiceEOuv, $this->token, $ultimaDataExecucao, $dataAtual, null, $idRelatorioImportacao);
+
+            //Caso retornado algum erro
+            if (is_string($retornoWs)) {
+                if (strpos($retornoWs, 'Invalidado') !== false) {
+                    //Tenta gerar novo token
+                    $tokenValido = MdCguEouvWS::apiValidarToken($this->urlWebServiceEOuv, $this->usuarioWebService, $this->senhaUsuarioWebService, $this->client_id, $this->client_secret);
+
+                    if (isset($tokenValido['error'])) {
+                        $textoMensagemErroToken = 'Não foi possível validar o Token de acesso aos WebServices do E-ouv. <br>Verifique as informações de Usuário, Senha, Client_Id e Client_Secret nas configurações de Parâmetros do Módulo';
+
+                    } elseif (isset($tokenValido['access_token'])) {
+                        $this->gravarParametroToken($tokenValido['access_token']);
+                        $token = $tokenValido['access_token'];
+
+                        //Chama novamente a execução da ConsultaManifestacao que deu errado por causa do Token
+                        $retornoWs = $this->executarServicoConsultaManifestacoes($this->urlWebServiceEOuv, $token, $ultimaDataExecucao, $dataAtual, null, $idRelatorioImportacao);
+                    }
+                }
+            }
+
+            if ($textoMensagemErroToken == '') {
+                $arrComErro = $this->obterManifestacoesComErro($this->urlWebServiceEOuv, $token, $ultimaDataExecucao, $dataAtual, $idRelatorioImportacao);
+
+                $arrManifestacoes = array();
+
+                if (is_array($retornoWs)) {
+                    // Filtra as manifestações e-Ouv
+                    $arrManifestacoes = array_filter($retornoWs, function($manifestacao) {
+                        return $manifestacao['TipoManifestacao']['IdTipoManifestacao'] <> 8;
+                    });
+                    $qtdManifestacoesNovas = count($arrManifestacoes);
+                }
+
+                if (is_array($arrComErro)) {
+                    $qtdManifestacoesAntigas = count($arrComErro);
+                    $arrManifestacoes = array_merge($arrManifestacoes, $arrComErro);
+                }
+
+                if (count($arrManifestacoes) > 0) {
+                    $semManifestacoesEncontradas = false;
+                    foreach ($arrManifestacoes as $retornoWsLinha) {
+                        $this->executarImportacaoLinha($retornoWsLinha);
+                    }
+                }
+
+                $textoMensagemFinal = 'Execução Finalizada com Sucesso!';
+                $SinSucessoExecucao = 'S';
+
+                if ($semManifestacoesEncontradas) {
+                    $textoMensagemFinal = $textoMensagemFinal . ' Não foram encontradas manifestações para o período.';
+                } else {
+                    $textoMensagemFinal = $textoMensagemFinal . '<br>Quantidade de Manifestações novas encontradas (e-Ouv): ' . $qtdManifestacoesNovas . '<br>Quantidade de Manifestações encontadas que ocorreram erro em outras importações: ' . $qtdManifestacoesAntigas;
+                }
+
+                if ($this->ocorreuErroEmProtocolo) {
+                    $textoMensagemFinal = $textoMensagemFinal . '<br> Ocorreram erros em 1 ou mais protocolos.';
+                }
+            } else {
+                $textoMensagemFinal = $textoMensagemErroToken;
+            }
+
+            //Grava a execução com sucesso se tiver corrido tudo bem
+            $objEouvRelatorioImportacaoDTO2 = new MdCguEouvRelatorioImportacaoDTO();
+
+            $objEouvRelatorioImportacaoDTO2->setNumIdRelatorioImportacao($objEouvRelatorioImportacaoDTO->getNumIdRelatorioImportacao());
+            $objEouvRelatorioImportacaoDTO2->setStrSinSucesso($SinSucessoExecucao);
+            $objEouvRelatorioImportacaoDTO2->setStrDeLogProcessamento($textoMensagemFinal);
+            $objEouvRelatorioImportacaoRN->alterar($objEouvRelatorioImportacaoDTO2);
+
+
+
+        } catch(Exception $e) {
+
+            $objEouvRelatorioImportacaoDTO3 = new MdCguEouvRelatorioImportacaoDTO();
+            $objEouvRelatorioImportacaoDTO3->setNumIdRelatorioImportacao($objEouvRelatorioImportacaoDTO->getNumIdRelatorioImportacao());
+            $strMensagem = 'Ocorreu um erro no processamento:' . $e;
+            $strMensagem = substr($strMensagem, 0, 500);
+            $objEouvRelatorioImportacaoDTO3->setStrDeLogProcessamento($strMensagem);
+
+            $objEouvRelatorioImportacaoRN->alterar($objEouvRelatorioImportacaoDTO3);
+
+            PaginaSEI::getInstance()->processarExcecao($e);
+        } finally {
+            //Restaura a sessão
+            SessaoSEI::getInstance()->setBolHabilitada($isBolHabilitada);
+        }
+    }
+
+    /**
+     * Função para importar as manifestações e-Sic do FalaBR (tipo 8)
+     */
+    public function executarImportacaoManifestacaoESic()
+    {
+
+        $debugLocal = false;
+
+        // Log
+        LogSEI::getInstance()->gravar('Rotina de Importação de Manifestações do FalaBR (e-Sic)', InfraLog::$INFORMACAO);
+
+        // Lista parâmetros
+        $objEouvParametroDTO = new MdCguEouvParametroDTO();
+        $objEouvParametroDTO->retTodos();
+
+        // Busca parâmetros do banco de dados da tabela md_eouv_parametros
+        $objEouvParametroRN = new MdCguEouvParametroRN();
+        $arrObjEouvParametroDTO = $objEouvParametroRN->listarParametro($objEouvParametroDTO);
+        $numRegistros = count($arrObjEouvParametroDTO);
+
+        $this->preencheVariaveis($numRegistros, $arrObjEouvParametroDTO);
+
+        /**
+         * Função para buscar o 'restante' do token sem o limite de 255 caracteres do SEI
+         */
+        $tokenPart2 = BancoSEI::getInstance()->consultarSql("select substring(de_valor_parametro, 256, 455) from md_eouv_parametros where no_parametro='TOKEN';")[0]['computed'];
+        $this->token = $this->token . $tokenPart2;
+
+        // Busca parâmetros do banco de dados da tabela infra_parametros
+        $objInfraParametro = new InfraParametro(BancoSEI::getInstance());
+        $idUsuarioSei = $objInfraParametro->getValor('ID_USUARIO_SEI');
+        $dataAtual = InfraData::getStrDataHoraAtual();
+
+        $isBolHabilitada = SessaoSEI::getInstance(false)->isBolHabilitada();
+        SessaoSEI::getInstance()->setBolHabilitada(false);
+
+        // Simula login inicial
+        self::simulaLogin($this->siglaSistema, $this->identificacaoServico, $this->idUnidadeEsicPrincipal);
+
+        // Executa a importação dos dados
+        try {
+
+            //Retorna dados da Última execução com Sucesso
+            $objUltimaExecucao = MdCguEouvAgendamentoINT::retornarUltimaExecucaoSucesso('R');
+
+            if ($objUltimaExecucao != null) {
+                // Debug Logs
+                $debugLocal && LogSEI::getInstance()->gravar('$objUltimaExecuxao (e-Sic):' . $objUltimaExecucao->getDthDthPeriodoFinal());
+
+                $ultimaDataExecucao = $objUltimaExecucao->getDthDthPeriodoFinal();
+            } else {
+                // Debug Logs
+                $debugLocal && LogSEI::getInstance()->gravar('$objUltimaExecuxao (e-Sic) é NULL');
+
+                //Primeira execução ou nenhuma executada com sucesso
+                $ultimaDataExecucao = $this->dataInicialImportacaoManifestacoes;
+            }
+
+            $semManifestacoesEncontradas = true;
+            $qtdManifestacoesNovas = 0;
+            $qtdManifestacoesAntigas = 0;
+            $semRecursosEncontrados = true;
+            $qtdRecursosNovos = 0;
+            $qtdRecursosAntigos = 0;
+
+            /**
+             * A função abaixo gravarLogImportacao recebe o tipo de manifestação 'R' (Recursos) para as manifestações do e-Sic
+             */
+
+            $objEouvRelatorioImportacaoDTO = $this->gravarLogImportacao($ultimaDataExecucao, $dataAtual, 'R');
+            $idRelatorioImportacao = $objEouvRelatorioImportacaoDTO->getNumIdRelatorioImportacao();
+            $objEouvRelatorioImportacaoRN = new MdCguEouvRelatorioImportacaoRN();
+            $SinSucessoExecucao = 'N';
+            $textoMensagemErroToken = '';
+
+
+            /**
+             * As funções abaixo fazem a busca no webservice dos dados a serem trabalhados na rotina de importação
+             */
+            $debugLocal && LogSEI::getInstance()->gravar('Iniciando a consulta inicial');
+            $retornoWs = $this->executarServicoConsultaManifestacoes($this->urlWebServiceEOuv, $this->token, $ultimaDataExecucao, $dataAtual, null, $idRelatorioImportacao);
+            $retornoWsRecursos = $this->executarServicoConsultaRecursos($this->urlWebServiceESicRecursos, $this->token, $ultimaDataExecucao, $dataAtual, null, $idRelatorioImportacao);
+
+            //Caso retornado algum erro - Manifestações e-Sic
+            if (is_string($retornoWs)) {
+                $debugLocal && LogSEI::getInstance()->gravar('Retorno da consulta $retornoWs é uma string: ' . $retornoWs);
+
+                if (strpos($retornoWs, 'Invalidado') !== false) {
+                    //Tenta gerar novo token
+                    $tokenValido = MdCguEouvWS::apiValidarToken($this->urlWebServiceEOuv, $this->usuarioWebService, $this->senhaUsuarioWebService, $this->client_id, $this->client_secret);
+
+                    if (isset($tokenValido['error'])) {
+                        $textoMensagemErroToken = 'Não foi possível validar o Token de acesso aos WebServices do E-ouv. <br>Verifique as informações de Usuário, Senha, Client_Id e Client_Secret nas configurações de Parâmetros do Módulo';
+
+                    } elseif (isset($tokenValido['access_token'])) {
+                        $this->gravarParametroToken($tokenValido['access_token']);
+                        $token = $tokenValido['access_token'];
+
+                        //Chama novamente a execução da ConsultaManifestacao que deu errado por causa do Token
+                        $retornoWs = $this->executarServicoConsultaManifestacoes($this->urlWebServiceEOuv, $token, $ultimaDataExecucao, $dataAtual, null, $idRelatorioImportacao);
+                        $retornoWsRecursos = $this->executarServicoConsultaRecursos($this->urlWebServiceESicRecursos, $token, $ultimaDataExecucao, $dataAtual, null, $idRelatorioImportacao);
+                    }
+                }
+            }
+
+            /**
+             * @todo - criar rotina para buscar recursos das manifestções com erro caso exista alguma na tabela de log
+             */
+
+            if ($textoMensagemErroToken == '') {
+
+                /**
+                 * @debug - manifestacao com erro
+                 * Comentar a linha abaixo para debugar um retorno manual
+                 */
+                $debugLocal && LogSEI::getInstance()->gravar('Inicia busca manifestação com erros');
+                $arrComErro = $this->obterManifestacoesComErro($this->urlWebServiceEOuv, $token, $ultimaDataExecucao, $dataAtual, $idRelatorioImportacao, 'R');
+
+                $arrManifestacoes = array();
+
+                if (is_array($retornoWs)) {
+                    // Filtra as manifestações e-Sic
+                    $arrManifestacoes = array_filter($retornoWs, function($manifestacao) {
+                        return $manifestacao['TipoManifestacao']['IdTipoManifestacao'] == 8;
+                    });
+                    $qtdManifestacoesNovas = count($arrManifestacoes);
+                    $debugLocal && LogSEI::getInstance()->gravar('Possui novas manifestações qtd: ' . $qtdManifestacoesNovas);
+
+                }
+
+                $arrRecursos = array();
+                if (isset($retornoWsRecursos) && is_array($retornoWsRecursos)) {
+                    $debugLocal && LogSEI::getInstance()->gravar('Possui recursos qtd: ' . count($retornoWsRecursos['Recursos']));
+                    $arrRecursos = $retornoWsRecursos['Recursos'];
+                    $qtdRecursosNovos = count($arrRecursos);
+
+                }
+
+                if (is_array($arrComErro)) {
+                    $debugLocal && LogSEI::getInstance()->gravar('Possui manifestações com erros - qtd: ' . count($arrComErro));
+                    $qtdManifestacoesAntigas = count($arrComErro);
+                    $arrManifestacoes = array_merge($arrManifestacoes, $arrComErro);
+
+                }
+
+                // Importa manifestações e-Sic
+                if (count($arrManifestacoes) > 0) {
+                    $semManifestacoesEncontradas = false;
+                    foreach ($arrManifestacoes as $retornoWsLinha) {
+                        $debugLocal && LogSEI::getInstance()->gravar('Inicia importação por Linha');
+                        $this->executarImportacaoLinha($retornoWsLinha, 'R');
+                    }
+                }
+                // Importa recursos e-Sic
+                if (count($arrRecursos) > 0) {
+                    $semRecursosEncontrados = false;
+                    foreach ($arrRecursos as $retornoWsLinha) {
+                        $debugLocal && LogSEI::getInstance()->gravar('Inicia importação por linha de Recursos - protocolo: ' . $retornoWsLinha['numProtocolo']);
+                        $this->executarImportacaoLinhaRecursos($retornoWsLinha);
+                    }
+                }
+
+                $textoMensagemFinal = 'Execução Finalizada com Sucesso!';
+                $SinSucessoExecucao = 'S';
+
+                if ($semManifestacoesEncontradas) {
+                    $textoMensagemFinal = $textoMensagemFinal . ' Não foram encontradas manifestações para o período.';
+                } else {
+                    $textoMensagemFinal = $textoMensagemFinal . '<br>Quantidade de Manifestações novas encontradas (e-Sic): ' . $qtdManifestacoesNovas . '<br>Quantidade de Manifestações encontadas que ocorreram erro em outras importações: ' . $qtdManifestacoesAntigas;
+                }
+
+                if ($semRecursosEncontrados) {
+                    $textoMensagemFinal = $textoMensagemFinal . ' Não foram encontrados recursos para o período.';
+                } else {
+                    $textoMensagemFinal = $textoMensagemFinal . '<br>Quantidade de Recursos novos encontrados (e-Sic): ' . $qtdRecursosNovos;
+                }
+
+                if ($this->ocorreuErroEmProtocolo) {
+                    $textoMensagemFinal = $textoMensagemFinal . '<br> Ocorreram erros em 1 ou mais protocolos.';
+                }
+            } else {
+                $textoMensagemFinal = $textoMensagemErroToken;
+            }
+
+            //Grava a execução com sucesso se tiver corrido tudo bem
+            $objEouvRelatorioImportacaoDTO2 = new MdCguEouvRelatorioImportacaoDTO();
+
+            $objEouvRelatorioImportacaoDTO2->setNumIdRelatorioImportacao($objEouvRelatorioImportacaoDTO->getNumIdRelatorioImportacao());
+            $objEouvRelatorioImportacaoDTO2->setStrSinSucesso($SinSucessoExecucao);
+            $objEouvRelatorioImportacaoDTO2->setStrDeLogProcessamento($textoMensagemFinal);
+            $objEouvRelatorioImportacaoDTO2->setStrTipManifestacao('R');
+            $objEouvRelatorioImportacaoRN->alterar($objEouvRelatorioImportacaoDTO2);
+
+            LogSEI::getInstance()->gravar('Finalizado a importção dos processos e-Sic - FalaBR');
+
+        } catch(Exception $e) {
+
+            $objEouvRelatorioImportacaoDTO3 = new MdCguEouvRelatorioImportacaoDTO();
+            $objEouvRelatorioImportacaoDTO3->setNumIdRelatorioImportacao($objEouvRelatorioImportacaoDTO->getNumIdRelatorioImportacao());
+            $strMensagem = 'Ocorreu um erro no processamento:' . $e;
+            $strMensagem = substr($strMensagem, 0, 500);
+            $objEouvRelatorioImportacaoDTO3->setStrDeLogProcessamento($strMensagem);
+            $objEouvRelatorioImportacaoRN->alterar($objEouvRelatorioImportacaoDTO3);
+
+            PaginaSEI::getInstance()->processarExcecao($e);
+
+        } finally {
+            //Restaura a Sessão
+            SessaoSEI::getInstance()->setBolHabilitada($isBolHabilitada);
+        }
     }
 
     public function preencheVariaveis(int $numRegistros, $arrObjEouvParametroDTO)
@@ -182,6 +543,235 @@ class MdCguEouvAgendamentoRN extends InfraRN
         }
     }
 
+    public function gravarParametroToken($tokenGerado){
+
+        $objEouvParametroDTO = new MdCguEouvParametroDTO();
+        $objEouvParametroDTO -> setNumIdParametro(10);
+        $objEouvParametroDTO -> setStrNoParametro('TOKEN');
+        $objEouvParametroDTO -> setStrDeValorParametro($tokenGerado);
+
+        $objEouvParametroRN = new MdCguEouvParametroRN();
+        $objEouvParametroRN->alterarParametro($objEouvParametroDTO);
+
+    }
+
+    public function executarServicoConsultaManifestacoes($urlConsultaManifestacao, $token, $ultimaDataExecucao, $dataAtual, $numprotocolo = null, $numIdRelatorio = null)
+    {
+        $arrParametrosUrl = array(
+            'dataCadastroInicio' => $ultimaDataExecucao,
+            'dataCadastroFim' => $dataAtual,
+            'numprotocolo' => $numprotocolo
+        );
+
+        $arrParametrosUrl = http_build_query($arrParametrosUrl);
+
+        $urlConsultaManifestacao = $urlConsultaManifestacao . "?" . $arrParametrosUrl;
+
+        $retornoWs = MdCguEouvWS::apiRestRequest($urlConsultaManifestacao, $token, 1);
+
+        if (is_null($numprotocolo)) {
+            // Verifica se retornou Token Invalido
+            if (is_string($retornoWs)) {
+                // Token expirado, necessário gerar novo Token
+                if (strpos($retornoWs, 'Invalidado') !== false) {
+                    return "Token Invalidado";
+                }
+                // Outro erro
+                if (strpos($retornoWs, 'Erro') !== false) {
+                    return "Erro:" . $retornoWs;
+                }
+            }
+        } else {
+            // Faz tratamento diferenciado para consulta por Protocolo específico
+            if(is_string($retornoWs)) {
+                if (strpos($retornoWs, '404') !== false) {
+                    $this->gravarLogLinha($this->formatarProcesso($numprotocolo), $numIdRelatorio, 'Nenhum retorno encontrado!', 'S');
+                    $retornoWs = null;
+                } elseif (strpos($retornoWs, 'Erro') !== false) {
+                    $this->gravarLogLinha($this->formatarProcesso($numprotocolo), $numIdRelatorio, "Erro desconhecido" . $retornoWs, 'N');
+                    throw new Exception($retornoWs);
+                }
+            }
+        }
+
+        return $retornoWs;
+    }
+
+    public function executarServicoConsultaRecursos($urlConsultaRecurso, $token, $ultimaDataExecucao = null, $dataAtual = null, $numprotocolo = null, $numIdRelatorio = null)
+    {
+
+        $debugLocal = false;
+        $debugLocal && LogSEI::getInstance()->gravar('[executarServicoConsultaRecursos] Parâmetros: $ultimaDataExecucao: ' . $ultimaDataExecucao . ' | $dataAtual: ' . $dataAtual . ' | $numprotocolo: ' . $numprotocolo);
+
+        $arrParametrosUrl = array(
+            'dataAberturaInicio' => $ultimaDataExecucao,
+            'dataAberturaFim' => $dataAtual,
+            'NumProtocolo' => $numprotocolo
+        );
+
+        $arrParametrosUrl = http_build_query($arrParametrosUrl);
+
+        $urlConsultaRecurso = $urlConsultaRecurso . "?" . $arrParametrosUrl;
+
+        $retornoWs = MdCguEouvWS::apiRestRequest($urlConsultaRecurso, $token, 1);
+
+        if (is_null($numprotocolo)) {
+            //Verifica se retornou Token Invalido
+            if (is_string($retornoWs)) {
+                if (strpos($retornoWs, 'Invalidado') !== false) {
+                    //Token expirado, necessÃ¡rio gerar novo Token
+                    return "Token Invalidado";
+                }
+
+                //Outro erro
+                if (strpos($retornoWs, 'Erro') !== false) {
+                    //Token expirado, necessÃ¡rio gerar novo Token
+                    return "Erro:" . $retornoWs;
+                }
+
+            }
+        } else {
+            //Faz tratamento diferenciado para consulta por Protocolo específico
+            if(is_string($retornoWs)) {
+                if (strpos($retornoWs, 'Erro') !== false) {
+                    if (strpos($retornoWs, '404') !== false) {
+                        $this->gravarLogLinha($this->formatarProcesso($numprotocolo), $numIdRelatorio, "Usuário não possui permissão de acesso neste protocolo.", 'N');
+                        $retornoWs = null;
+                    } else {
+                        $this->gravarLogLinha($this->formatarProcesso($numprotocolo), $numIdRelatorio, "Erro desconhecido" . $retornoWs, 'N');
+                        throw new Exception($retornoWs);
+                    }
+                }
+            }
+        }
+
+        return $retornoWs;
+    }
+
+    public function criarNovoProcesso($idTipoManifestacaoSei, $idTipoManifestacao, bool $manifestacaoESic, $arrDetalheManifestacao, $idUnidadeDestino, string $numProtocoloFormatado, $tipoManifestacao, $arrRecursosManifestacao)
+    {
+        /**
+         * Inicia criação do Procedimento de criação de novo Processo
+         */
+        try {
+            $objTipoProcedimentoDTO = new TipoProcedimentoDTO();
+            $objTipoProcedimentoDTO->retNumIdTipoProcedimento();
+            $objTipoProcedimentoDTO->retStrNome();
+            $objTipoProcedimentoDTO->retStrStaNivelAcessoSugestao();
+            $objTipoProcedimentoDTO->retStrStaGrauSigiloSugestao();
+            $objTipoProcedimentoDTO->retStrSinIndividual();
+            $objTipoProcedimentoDTO->retNumIdHipoteseLegalSugestao();
+            $objTipoProcedimentoDTO->setNumIdTipoProcedimento($idTipoManifestacaoSei);
+
+            $objTipoProcedimentoRN = new TipoProcedimentoRN();
+            $objTipoProcedimentoDTO = $objTipoProcedimentoRN->consultarRN0267($objTipoProcedimentoDTO);
+
+            if ($objTipoProcedimentoDTO == null) {
+                throw new Exception('Tipo de processo não encontrado: ' . $idTipoManifestacaoSei);
+            }
+
+            /*
+             * Verifica se deve importar documentos tipo 8 (e-Sic)
+             */
+            if ($idTipoManifestacao == 8 && !$manifestacaoESic) {
+                return;
+            }
+
+            $objProcedimentoAPI = new ProcedimentoAPI();
+            $objProcedimentoAPI->setIdTipoProcedimento($objTipoProcedimentoDTO->getNumIdTipoProcedimento());
+
+            $varEspecificacaoAssunto = "";
+
+            if (is_array($arrDetalheManifestacao['Assunto'])) {
+                $varEspecificacaoAssunto = $arrDetalheManifestacao['Assunto']['DescAssunto'];
+            }
+            if (is_array($arrDetalheManifestacao['SubAssunto'])) {
+                $varEspecificacaoAssunto = $varEspecificacaoAssunto . " / " . $arrDetalheManifestacao['SubAssunto']['DescSubAssunto'];
+            }
+
+            $objProcedimentoAPI->setEspecificacao($varEspecificacaoAssunto);
+            $objProcedimentoAPI->setIdUnidadeGeradora($idUnidadeDestino);
+            $objProcedimentoAPI->setNumeroProtocolo($numProtocoloFormatado);
+            $objProcedimentoAPI->setDataAutuacao($arrDetalheManifestacao['DataCadastro']);
+            $objProcedimentoAPI->setNivelAcesso($objTipoProcedimentoDTO->getStrStaNivelAcessoSugestao());
+            $objProcedimentoAPI->setGrauSigilo($objTipoProcedimentoDTO->getStrStaGrauSigiloSugestao());
+            $objProcedimentoAPI->setIdHipoteseLegal($objTipoProcedimentoDTO->getNumIdHipoteseLegalSugestao());
+            $objProcedimentoAPI->setObservacao("Processo Gerado Automaticamente pela Integração SEI x FalaBR");
+
+            $objEntradaGerarProcedimentoAPI = new EntradaGerarProcedimentoAPI();
+            $objEntradaGerarProcedimentoAPI->setProcedimento($objProcedimentoAPI);
+
+            $objSaidaGerarProcedimentoAPI = new SaidaGerarProcedimentoAPI();
+
+            $objSeiRN = new SeiRN();
+
+            $arrDocumentos = $this->gerarAnexosProtocolo($arrDetalheManifestacao['Teor']['Anexos'], $numProtocoloFormatado, $tipoManifestacao);
+
+            /**
+             * Verificar o tipo de documento a ser importado para gerar o PDF conforme tipo de documento
+             */
+            if ($manifestacaoESic) {
+                $documentoManifestacao = $this->gerarPDFDocumentoESic($arrDetalheManifestacao, $arrRecursosManifestacao);
+            } else {
+                $documentoManifestacao = $this->gerarPDFPedidoInicial($arrDetalheManifestacao);
+            }
+
+            LogSEI::getInstance()->gravar('Importação de Manifestação ' . $numProtocoloFormatado . ': total de  Anexos configurados: ' . count($arrDocumentos), InfraLog::$INFORMACAO);
+
+            array_unshift($arrDocumentos, $documentoManifestacao);
+            $objEntradaGerarProcedimentoAPI->setDocumentos($arrDocumentos);
+            $objSaidaGerarProcedimentoAPI = $objSeiRN->gerarProcedimento($objEntradaGerarProcedimentoAPI);
+            $this->gravarLogLinha($numProtocoloFormatado, $this->idRelatorioImportacao, 'Protocolo ' . $arrDetalheManifestacao['numProtocolo'] . ' gravado com sucesso.', 'S', $tipoManifestacao);
+
+        } catch (Exception $e) {
+
+            if ($objSaidaGerarProcedimentoAPI != null and $objSaidaGerarProcedimentoAPI->getIdProcedimento() > 0) {
+                $this->excluirProcessoComErro($objSaidaGerarProcedimentoAPI->getIdProcedimento());
+            }
+            $this->gravarLogLinha($numProtocoloFormatado, $this->idRelatorioImportacao, 'Erro na gravação: ' . $e, 'N', $tipoManifestacao);
+        }
+        return $arrDocumentos;
+    }
+
+    public function getUnidadeDestino(string $tipo_recurso)
+    {
+        // Vincular Recursos com as unidades corretas conforme o tipo de recurso
+        // Se for 1 instância envia processo para ESIC_ID_UNIDADE_RECURSO_PRIMEIRA_INSTANCIA
+        if ($tipo_recurso == 'R1') {
+            $unidadeDestino = $this->idUnidadeRecursoPrimeiraInstancia;
+        } elseif ($tipo_recurso == 'R2') {
+            $unidadeDestino = $this->idUnidadeRecursoSegundaInstancia;
+        } elseif ($tipo_recurso == 'R3' || $tipo_recurso == 'RC') {
+            $unidadeDestino = $this->idUnidadeRecursoTerceiraInstancia;
+        } elseif ($tipo_recurso == 'PR') {
+            $unidadeDestino = $this->idUnidadeRecursoPedidoRevisao;
+        } else {
+            $unidadeDestino = $this->idUnidadeOuvidoria;
+        }
+        return $unidadeDestino;
+    }
+
+    // GZIP DECODE
+    function gzdecode($data)
+    {
+
+        return gzinflate(substr($data, 10, -8));
+    }
+
+    public function verificaRetornoWS($retornoWsLista)
+    {
+        /*
+        função criada para tratar o retorno de dados do WS, pois quando existe apenas um unico resultado retorna um objeto e
+        quando tem mais de um resultado retorna um array ocasionando falhas na exibição dos dados.
+        */
+        if (isset($retornoWsLista) and key_exists(0, $retornoWsLista)) {
+            $resultado = $retornoWsLista;
+        } else {
+            $resultado = array ( $retornoWsLista );
+        }
+        return $resultado;
+    }
+
     public function gravarLogImportacao($ultimaDataExecucao, $dataAtual, $tipoManifestacao = 'P'){
 
         try {
@@ -209,358 +799,868 @@ class MdCguEouvAgendamentoRN extends InfraRN
     }
 
     /**
-     * Função para importar as manifestações e-Ouv do FalaBR
-     *
-     * Tipos: 1, 2, 3, 4, 5, 6 e 7
+     * Grava log detalhado
      */
-    public function executarImportacaoManifestacaoEOuv()
+    public function gravarLogLinha($numProtocolo, $idRelatorioImportacao, $mensagem, $sinSucesso, $tipoManifestacao = 'P', $dataPrazoAtendimento = null)
     {
-        // Log
-        LogSEI::getInstance()->gravar('Rotina de Importação de Manifestações do E-Ouv', InfraLog::$INFORMACAO);
+        $objEouvRelatorioImportacaoDetalheDTO = new MdCguEouvRelatorioImportacaoDetalheDTO();
+        $objEouvRelatorioImportacaoDetalheDTO->retStrProtocoloFormatado();
+        $objEouvRelatorioImportacaoDetalheDTO->setNumIdRelatorioImportacao($idRelatorioImportacao);
+        $objEouvRelatorioImportacaoDetalheDTO->setStrProtocoloFormatado($numProtocolo);
 
-        $objEouvParametroDTO = new MdCguEouvParametroDTO();
-        $objEouvParametroDTO->retTodos();
+        $objEouvRelatorioImportacaoDetalheRN = new MdCguEouvRelatorioImportacaoDetalheRN();
+        $objExisteDetalheDTO = $objEouvRelatorioImportacaoDetalheRN->consultar($objEouvRelatorioImportacaoDetalheDTO);
 
-        // Busca parâmetros do banco de dados
-        $objEouvParametroRN = new MdCguEouvParametroRN();
-        $arrObjEouvParametroDTO = $objEouvParametroRN->listarParametro($objEouvParametroDTO);
+        $objEouvRelatorioImportacaoDetalheDTO->setStrSinSucesso($sinSucesso);
+        $objEouvRelatorioImportacaoDetalheDTO->setStrTipManifestacao($tipoManifestacao);
+        $objEouvRelatorioImportacaoDetalheDTO->setStrDescricaoLog(substr($mensagem,0,254));
+        $objEouvRelatorioImportacaoDetalheDTO->setDthDthImportacao(InfraData::getStrDataHoraAtual());
+        $objEouvRelatorioImportacaoDetalheDTO->setDthDthPrazoAtendimento($dataPrazoAtendimento);
 
-        $numRegistros = count($arrObjEouvParametroDTO);
+        if ($objExisteDetalheDTO==null) {
+            $objEouvRelatorioImportacaoDetalheRN->cadastrar($objEouvRelatorioImportacaoDetalheDTO);
+        } else {
+            $objEouvRelatorioImportacaoDetalheRN->alterar($objEouvRelatorioImportacaoDetalheDTO);
+        }
+    }
 
-        $this->preencheVariaveisEouv($numRegistros, $arrObjEouvParametroDTO);
+    public function obterManifestacoesComErro($urlConsultaManifestacao, $token, $ultimaDataExecucao, $dataAtual, $numIdRelatorio, $TipoManifestacao = 'P')
+    {
+        $objEouvRelatorioImportacaoDetalheDTO = new MdCguEouvRelatorioImportacaoDetalheDTO();
+        $objEouvRelatorioImportacaoDetalheDTO->retStrProtocoloFormatado();
+        $objEouvRelatorioImportacaoDetalheDTO->setStrSinSucesso('N');
+        $objEouvRelatorioImportacaoDetalheDTO->setStrTipManifestacao($TipoManifestacao);
+
+        $objEouvRelatorioImportacaoDetalheRN = new MdCguEouvRelatorioImportacaoDetalheRN();
+        $objListaErros = $objEouvRelatorioImportacaoDetalheRN->listar($objEouvRelatorioImportacaoDetalheDTO);
+
+        $arrResult = array();
+        $arrProtocolos = array();
+
+        foreach($objListaErros as $erro) {
+
+            $numProtocolo = preg_replace("/[^0-9]/", "", $erro->getStrProtocoloFormatado());
+
+            //Se já estiver na lista não faz novamente para determinado protocolo
+            if (!in_array($numProtocolo, $arrProtocolos)){
+
+                //Adiciona no array de Protocolos
+                array_push($arrProtocolos, $numProtocolo);
+
+                $retornoWsErro = $this->executarServicoConsultaManifestacoes($urlConsultaManifestacao, $token, null, $dataAtual, $numProtocolo, $numIdRelatorio);
+
+                if (!is_null($retornoWsErro) && $retornoWsErro <> ''){
+                    $arrResult = array_merge($arrResult, $retornoWsErro);
+                }
+            }
+        }
+
+        return $arrResult;
+    }
+
+    private static function obterServico($SiglaSistema, $IdentificacaoServico){
+
+        $objUsuarioDTO = new UsuarioDTO();
+        $objUsuarioDTO->retNumIdUsuario();
+        $objUsuarioDTO->setStrSigla($SiglaSistema);
+        $objUsuarioDTO->setStrStaTipo(UsuarioRN::$TU_SISTEMA);
+
+        $objUsuarioRN = new UsuarioRN();
+        $objUsuarioDTO = $objUsuarioRN->consultarRN0489($objUsuarioDTO);
+
+        if ($objUsuarioDTO==null){
+            throw new InfraException('Sistema ['.$SiglaSistema.'] não encontrado.');
+        }
+
+        $objServicoDTO = new ServicoDTO();
+        $objServicoDTO->retNumIdServico();
+        $objServicoDTO->retStrIdentificacao();
+        $objServicoDTO->retStrSiglaUsuario();
+        $objServicoDTO->retNumIdUsuario();
+        $objServicoDTO->retStrServidor();
+        $objServicoDTO->retStrSinLinkExterno();
+        $objServicoDTO->retNumIdContatoUsuario();
+        $objServicoDTO->setNumIdUsuario($objUsuarioDTO->getNumIdUsuario());
+        $objServicoDTO->setStrIdentificacao($IdentificacaoServico);
+
+        $objServicoRN = new ServicoRN();
+        $objServicoDTO = $objServicoRN->consultar($objServicoDTO);
+
+        if ($objServicoDTO==null){
+            throw new InfraException('Serviço ['.$IdentificacaoServico.'] do sistema ['.$SiglaSistema.'] não encontrado.');
+        }
+
+        return $objServicoDTO;
+    }
+
+    private static function obterUnidade($IdUnidade, $SiglaUnidade){
+
+        $objUnidadeDTO = new UnidadeDTO();
+        $objUnidadeDTO->retNumIdUnidade();
+        $objUnidadeDTO->retStrSigla();
+        $objUnidadeDTO->retStrDescricao();
+
+        if($IdUnidade!=null) {
+            $objUnidadeDTO->setNumIdUnidade($IdUnidade);
+        }
+        if($SiglaUnidade!=null){
+            $objUnidadeDTO->setStrSigla($SiglaUnidade);
+        }
+
+        $objUnidadeRN = new UnidadeRN();
+        $objUnidadeDTO = $objUnidadeRN->consultarRN0125($objUnidadeDTO);
+
+        if ($objUnidadeDTO==null){
+            throw new InfraException('Unidade ['.$IdUnidade.'] não encontrada.');
+        }
+
+        return $objUnidadeDTO;
+    }
+
+    public function executarImportacaoLinha($retornoWsLinha, $tipoManifestacao = 'P')
+    {
+        $debugLocal = false;
+
+        $objProcedimentoDTO = new ProcedimentoDTO();
+        $objProcedimentoDTO->setDblIdProcedimento(null);
+
+        $linkDetalheManifestacao = $retornoWsLinha['Links'][0]['href'];
+        $arrDetalheManifestacao = MdCguEouvWS::apiRestRequest($linkDetalheManifestacao, $this->token, 2);
 
         /**
-         * Função para buscar o 'restante' do token sem o limite de 255 caracteres do SEI
+         * Verifica Tipo de Manifestação e-Ouv ou e-Sic
          */
-        $tokenPart2 = BancoSEI::getInstance()->consultarSql('select substring(de_valor_parametro, 256, 455) from md_eouv_parametros where id_parametro=10;')[0]['computed'];
-        $this->token = $this->token . $tokenPart2;
+        if ($tipoManifestacao == 'P' && $retornoWsLinha['TipoManifestacao']['IdTipoManifestacao'] <> 8) {
+            $debugLocal && LogSEI::getInstance()->gravar('Importação tipo "P" - tipoManifestação <> "8"');
+            $manifestacaoESic = false;
+            $idUnidadeDestino = $this->idUnidadeOuvidoria;
+        } elseif ($tipoManifestacao == 'R' && $retornoWsLinha['TipoManifestacao']['IdTipoManifestacao'] == 8) {
+            $debugLocal && LogSEI::getInstance()->gravar('Importação tipo "R" - tipoManifestação == "8"');
+            $manifestacaoESic = true;
+            $idUnidadeDestino = $this->idUnidadeEsicPrincipal;
 
-        $dataAtual = InfraData::getStrDataHoraAtual();
+            /**
+             * Importar Recursos caso seja manifestação e-Sic (Tipo 8)
+             */
+            $arrRecursosManifestacao = MdCguEouvWS::apiRestRequest($this->urlWebServiceESicRecursos . '?NumProtocolo=' . $arrDetalheManifestacao['NumerosProtocolo'][0], $this->token, 2);
+        }
 
-        $isBolHabilitada = SessaoSEI::getInstance(false)->isBolHabilitada();
-        SessaoSEI::getInstance()->setBolHabilitada(false);
+        $numProtocoloFormatado =  $this->formatarProcesso($arrDetalheManifestacao['NumerosProtocolo'][0]);
 
-        $mdCguEouvWS = new MdCguEouvWS($this->urlWebServiceEOuv, $this->urlWebServiceESicRecursos, $this->idTipoDocumentoAnexoDadosManifestacao,
-            $this->idUnidadeOuvidoria, $this->idUnidadeEsicPrincipal, $this->idUnidadeRecursoPrimeiraInstancia, $this->idUnidadeRecursoSegundaInstancia,
-            $this->idUnidadeRecursoTerceiraInstancia, $this->idUnidadeRecursoPedidoRevisao, $this->token);
-        // Simula login inicial
-        $mdCguEouvWS->simulaLogin($this->siglaSistema, $this->identificacaoServico, $this->idUnidadeOuvidoria);
+        // Verifica se o tipo de manifestação é suportado
+        $numIdTipoManifestacao = $retornoWsLinha['TipoManifestacao']['IdTipoManifestacao'];
+        if ($numIdTipoManifestacao > 8) {
+            // Se não for marca como sucesso para evitar reimportação na próxima execução.
+            $this->gravarLogLinha($numProtocoloFormatado, $this->idRelatorioImportacao,
+                'Tipo de manifestação não suportado (ID = '.$numIdTipoManifestacao.'). Não será importada.', 'S');
+            return;
+        }
 
+
+        /**
+         * Esta data é gravada na tabela de log detalhada
+         * Em caso de alteração no prazo do atendimento será feita nova importação dos dados do recurso
+         * Verifica se o retorno dos recursos não é uma string
+         */
+        if ($arrRecursosManifestacao <> '' && !is_string($arrRecursosManifestacao)) {
+            $debugLocal && LogSEI::getInstance()->gravar('Possui $arrRecursosManifestacao - qtd: ' . count($arrRecursosManifestacao['Recursos']));
+            $dataPrazoAtendimento = $arrRecursosManifestacao['Recursos'][(count($arrRecursosManifestacao['Recursos']) - 1)]['prazoAtendimento'];
+        } else {
+            $debugLocal && LogSEI::getInstance()->gravar('NÃO possui $arrRecursosManifestacao');
+            $dataPrazoAtendimento = $retornoWsLinha['PrazoAtendimento'];
+        }
+
+        /**
+         * Limpa os registros de detalhe de importação com erro para este NUP.
+         * Caso ocorra um novo, será criado novo registro de erro para o NUP no tratamento desta function.
+         */
+        $this->limparErrosParaNup($numProtocoloFormatado);
+
+        if (!isset($arrDetalheManifestacao['TipoManifestacao']['IdTipoManifestacao'])) {
+            $this->gravarLogLinha($numProtocoloFormatado, $this->idRelatorioImportacao, 'Tipo de processo não foi informado.', 'N');
+            /**
+             * @todo - não deveria parara aqui? se não tiver um tipo de processo não informado?
+             */
+        } else {
+            $objEouvDeparaImportacaoDTO = new MdCguEouvDeparaImportacaoDTO();
+            $objEouvDeparaImportacaoDTO->retNumIdTipoProcedimento();
+            $objEouvDeparaImportacaoDTO->setNumIdTipoManifestacaoEouv($arrDetalheManifestacao['TipoManifestacao']['IdTipoManifestacao']);
+
+            $objEouvDeparaImportacaoRN = new MdCguEouvDeparaImportacaoRN();
+            $objEouvDeparaImportacaoDTO = $objEouvDeparaImportacaoRN->consultar($objEouvDeparaImportacaoDTO);
+
+            if (!$objEouvDeparaImportacaoDTO == null) {
+                $idTipoManifestacaoSei = $objEouvDeparaImportacaoDTO->getNumIdTipoProcedimento();
+            } else {
+                $this->gravarLogLinha($numProtocoloFormatado, $this->objEouvRelatorioImportacaoDTO->getNumIdRelatorioImportacao(), 'Não existe mapeamento DePara do Tipo de Manifestação do FalaBR (E-Ouv|E-Sic) para o tipo de procedimento do SEI.', 'N');
+                //continue;
+            }
+        }
+
+        /**
+         * Se for Manifestação do e-Sic verificar:houve alteração na data 'PrazoAtendimento' e
+         * gera novo arquivo PDF com as alterações para inserção no mesmo protocolo (NUP) e
+         * importa anexos comparando o hash do arquivo para não duplicidade no processo
+         */
+        // Vefificar se o NUP já existe
+        $objProtocoloDTOExistente = $this->verificarProtocoloExistente($this->formatarProcesso($numProtocoloFormatado));
+
+        // 1. Caso já exista um Protocolo no SEI com o mesmo NUP
+        if (! is_null($objProtocoloDTOExistente)) {
+            // 2. Se existir e for e-Ouv
+            if ($tipoManifestacao == 'P') {
+                $debugLocal && LogSEI::getInstance()->gravar('Importando Linha Manifestação e-ouv - protocolo: ' . $this->formatarProcesso($numProtocoloFormatado));
+                // 2.1 Importar anexos novos se existirem... e retornar log
+                // @todo - melhoria próxima versão e-Ouv
+                $this->gravarLogLinha($numProtocoloFormatado, $this->idRelatorioImportacao, 'Erro na gravação: ' . 'Já existe um processo (e-Ouv) utilizando o número de protocolo.', 'N', $tipoManifestacao);
+            }
+
+            // 3. Se existir e for e-Sic
+            if ($tipoManifestacao == 'R') {
+
+                $debugLocal && LogSEI::getInstance()->gravar('Importando Linha Manifestação e-SIC - protocolo: ' . $this->formatarProcesso($numProtocoloFormatado));
+
+                /**
+                 * @todo - @teste
+                 * Teste aqui pra validar se o prazo sendo 'maior' na petição inicial já não deve importar os recursos..... (??)
+                 */
+                // Data do último prazo de atendimento para este protocolo
+                $objUltimaDataPrazoAtendimento = MdCguEouvAgendamentoINT::retornarUltimaDataPrazoAtendimento($numProtocoloFormatado);
+
+
+                // 4. Verificar se houve alteração na data 'PrazoAtendimento'
+                if (isset($objUltimaDataPrazoAtendimento) && $objUltimaDataPrazoAtendimento->getDthDthPrazoAtendimento() > $dataPrazoAtendimento) {
+
+                    // Importar anexos do novo recurso
+                    try {
+                        $anexoCount = 0;
+                        if (isset($arrRecursosManifestacao['Recursos']) && is_array($arrRecursosManifestacao['Recursos'])) {
+
+                            // Verifica Tipo de Recurso
+                            $tipo_recurso = $this->verificaTipo($arrRecursosManifestacao['Recursos']);
+
+                            $debugLocal && LogSEI::getInstance()->gravar('[executarImportacaoLinha] Importando o recurso do protocolo: ' . $numProtocoloFormatado);
+
+                            // Carregar documento recurso
+                            $this->gerarPDFDocumentoESic($arrDetalheManifestacao, $arrRecursosManifestacao, $objProtocoloDTOExistente->getDblIdProtocolo(), $tipo_recurso);
+                            LogSEI::getInstance()->gravar('Módulo Integração FalaBR - Importação de Recurso ' . $numProtocoloFormatado . ': total de  Anexos configurados: ' . $anexoCount, InfraLog::$INFORMACAO);
+                            $this->gravarLogLinha($numProtocoloFormatado, $this->idRelatorioImportacao, 'Recurso com protocolo ' . $numProtocoloFormatado . ' importado com sucesso com ' . $anexoCount . ' anexos incluidos no protocolo.', 'S', $tipoManifestacao, $dataPrazoAtendimento);
+
+                            // Carregar anexos
+                            $recursos = $arrRecursosManifestacao['Recursos'];
+                            foreach ($recursos as $recurso) {
+                                if (count($recurso['anexos']) > 0) {
+                                    $anexosAdicionados = $this->gerarAnexosProtocolo($recurso['anexos'], $numProtocoloFormatado, $tipoManifestacao, $objProtocoloDTOExistente->getDblIdProtocolo());
+                                    if (count($anexosAdicionados) > 0) {
+                                        $anexoCount++;
+                                    }
+                                }
+                            }
+
+                            // Vincular Recursos com as unidades corretas conforme o tipo de recurso
+                            // Se for 1 instância envia processo para ESIC_ID_UNIDADE_RECURSO_PRIMEIRA_INSTANCIA
+                            if ($tipo_recurso == 'R1') {
+                                $unidadeDestino = $this->idUnidadeRecursoPrimeiraInstancia;
+                            } elseif ($tipo_recurso == 'R2') {
+                                $unidadeDestino = $this->idUnidadeRecursoSegundaInstancia;
+                            } elseif ($tipo_recurso == 'R3' || $tipo_recurso == 'RC') {
+                                $unidadeDestino = $this->idUnidadeRecursoTerceiraInstancia;
+                            } elseif ($tipo_recurso == 'PR') {
+                                $unidadeDestino = $this->idUnidadeRecursoPedidoRevisao;
+                            } else {
+                                $unidadeDestino = $this->idUnidadeOuvidoria;
+                            }
+
+                            try {
+                                $objEntradaEnviarProcesso = new EntradaEnviarProcessoAPI();
+                                $objEntradaEnviarProcesso->setIdProcedimento($objProtocoloDTOExistente->getDblIdProtocolo());
+                                $objEntradaEnviarProcesso->setUnidadesDestino([$unidadeDestino]);
+                                $objEntradaEnviarProcesso->setSinManterAbertoUnidade('S');
+                                $objEntradaEnviarProcesso->setSinEnviarEmailNotificacao('S');
+                                $objEntradaEnviarProcesso->setSinReabrir('S');
+
+                                $objSeiRN = new SeiRN();
+                                $objSeiRN->enviarProcesso($objEntradaEnviarProcesso);
+                                LogSEI::getInstance()->gravar('Módulo Integração FalaBR - (Recurso tipo ' . $tipo_recurso . ') Processo ' . $numProtocoloFormatado . ' enviado para unidade ' . $this->idUnidadeRecursoPrimeiraInstancia, InfraLog::$INFORMACAO);
+
+                            } catch (Exception $e) {
+                                LogSEI::getInstance()->gravar('Módulo Integração FalaBR - (Recurso tipo ' . $tipo_recurso . ') Não foi possivel abrir o Processo ' . $numProtocoloFormatado . ' na unidade ' . $this->idUnidadeRecursoPrimeiraInstancia . ' - erro: ' . $e, InfraLog::$INFORMACAO);
+                            }
+                        } else {
+                            /**
+                             * @todo - confirmar - aqui deve ficar como 'N' ou 'S'? Se fircar como 'N' entra como erro... ?? e é preciso gravar que não houve recurso mas teve alteração na data de prazo de atencimento,
+                             * esta data precisa ser salva no banco de dados... comentar/documentar aqui!
+                             */
+                            $this->gravarLogLinha($numProtocoloFormatado, $this->idRelatorioImportacao, 'Sem recursos novos.', 'S', $tipoManifestacao, $dataPrazoAtendimento);
+                        }
+                    } catch (Exception $e) {
+                        $this->gravarLogLinha($numProtocoloFormatado, $this->idRelatorioImportacao, 'Erro na gravação recurso: ' . $e, 'N', $tipoManifestacao);
+                    }
+                } else {
+                    // 4.2 Se não houve alteração na data 'PrazoAtendimento' retornar log
+                    $this->gravarLogLinha($numProtocoloFormatado, $this->idRelatorioImportacao, 'Já existe um processo (e-Sic) utilizando o número de protocolo e não há alteração para nova importação.', 'S', $tipoManifestacao, $dataPrazoAtendimento);
+                }
+            }
+        } else {
+            $this->criarNovoProcesso($idTipoManifestacaoSei, $retornoWsLinha['TipoManifestacao']['IdTipoManifestacao'], $manifestacaoESic, $arrDetalheManifestacao, $idUnidadeDestino, $numProtocoloFormatado, $tipoManifestacao, $arrRecursosManifestacao);
+        }
+    }
+
+    public function executarImportacaoLinhaRecursos ($arrRecursosManifestacao, $tipoManifestacao = 'R')
+    {
+        $debugLocal = false;
+
+        $objProcedimentoDTO = new ProcedimentoDTO();
+        $objProcedimentoDTO->setDblIdProcedimento(null);
+
+        $numProtocoloFormatado =  $this->formatarProcesso($arrRecursosManifestacao['numProtocolo']);
+        $dataPrazoAtendimento = $arrRecursosManifestacao['prazoAtendimento'];
+
+        /**
+         * Limpa os registros de detalhe de importação com erro para este NUP.
+         * Caso ocorra um novo, será criado novo registro de erro para o NUP no tratamento desta function.
+         */
+        $this->limparErrosParaNup($numProtocoloFormatado);
+
+        /**
+         * Se for Manifestação do e-Sic verificar:houve alteração na data 'PrazoAtendimento' e
+         * gera novo arquivo PDF com as alterações para inserção no mesmo protocolo (NUP) e
+         * importa anexos comparando o hash do arquivo para não duplicidade no processo
+         */
+        // Vefificar se o NUP já existe
+        $objProtocoloDTOExistente = $this->verificarProtocoloExistente($numProtocoloFormatado);
+
+        // Caso já exista um Protocolo no SEI continua, caso contrário apenas registra o log
+        if (! is_null($objProtocoloDTOExistente)) {
+
+            $debugLocal && LogSEI::getInstance()->gravar('[executarImportacaoLinhaRecursos] Existe o protocolo: ' . $numProtocoloFormatado);
+
+            // Se existir e for e-Sic
+            if ($tipoManifestacao == 'R') {
+
+                $debugLocal && LogSEI::getInstance()->gravar('[executarImportacaoLinhaRecursos] É do tipo: ' . $tipoManifestacao . ' > ' . $this->verificaTipo($arrRecursosManifestacao, 'R'));
+
+                // Data do último prazo de atendimento para este protocolo sem o tipo de recurso para buscar qualquer um recurso anterior
+                $objUltimaDataPrazoAtendimento = MdCguEouvAgendamentoINT::retornarUltimaDataPrazoAtendimento($numProtocoloFormatado);
+                $debugLocal && LogSEI::getInstance()->gravar('[executarImportacaoLinhaRecursos] Último prazo de atendimento: ' . $objUltimaDataPrazoAtendimento);
+
+                /**
+                 * Regra de bloqueio na criação de novos recursos caso já exista um recurso superior ao atualmente listado
+                 * - regra implementada devido à duplicidade na importação dos processos
+                 */
+                $ultimoTipoRecursoImportado = MdCguEouvAgendamentoINT::retornarTipoManifestacao($this->idRelatorioImportacao, $numProtocoloFormatado);
+                $ultimoTipoRecursoImportado = $ultimoTipoRecursoImportado ? $ultimoTipoRecursoImportado->getStrTipManifestacao() : $this->verificaTipo($arrRecursosManifestacao, 'R1');
+                $debugLocal && LogSEI::getInstance()->gravar('[executarImportacaoLinhaRecursos] Ultimo tipo de recurso importado: ' . $ultimoTipoRecursoImportado . ' - Tipo recurso atual: ' . $this->verificaTipo($arrRecursosManifestacao, 'R1'));
+
+                $permiteImportacaoRecursoAtual = $this->permiteImportacaoRecursoAtual($this->verificaTipo($arrRecursosManifestacao, 'R1'), $ultimoTipoRecursoImportado);
+                $debugLocal && LogSEI::getInstance()->gravar('[executarImportacaoLinhaRecursos] Permite criar o recurso atual: ' . $permiteImportacaoRecursoAtual);
+
+                if ($permiteImportacaoRecursoAtual == 'bloquear') {
+                    $debugLocal && LogSEI::getInstance()->gravar('[executarImportacaoLinhaRecursos] Não foi permitido criar o recurso, pode deve haver recurso anterior já importado');
+                    // Se não for permitido criar o recurso
+                    $this->gravarLogLinha($numProtocoloFormatado, $this->idRelatorioImportacao, 'O recurso existente no FalaBR não será importado devido à regra implementada - tipoAtual: "' . $this->verificaTipo($arrRecursosManifestacao, 'R') . '" | tipoAnterior: '. $ultimoTipoRecursoImportado .' | protocolo.', 'S', $ultimoTipoRecursoImportado, $objUltimaDataPrazoAtendimento->getDthDthPrazoAtendimento());
+                    return;
+                }
+
+                // Verificar se houve alteração na data 'PrazoAtendimento'
+                if (($objUltimaDataPrazoAtendimento && $objUltimaDataPrazoAtendimento->getDthDthPrazoAtendimento() <> $dataPrazoAtendimento) || $objUltimaDataPrazoAtendimento === null) {
+
+                    $debugLocal && LogSEI::getInstance()->gravar('[executarImportacaoLinhaRecursos] Data de prazo de atendimento diferente da última, incinia importacao');
+
+                    // Importar anexos do novo recurso
+                    try {
+                        if (isset($arrRecursosManifestacao)) {
+                            $anexoCount = isset($arrRecursosManifestacao['qtdAnexos']) ? $arrRecursosManifestacao['qtdAnexos'] : 0;
+
+                            // Verifica Tipo de Recurso
+                            $tipo_recurso = $this->verificaTipo($arrRecursosManifestacao);
+                            $unidadeDestino = $this->getUnidadeDestino($tipo_recurso);
+
+                            $debugLocal && LogSEI::getInstance()->gravar('[executarImportacaoLinhaRecursos] Tipo de recurso: ' . $tipo_recurso);
+
+                            // Buscar dados da Manifestação
+                            $numProtocoloSemFormatacao = str_replace(['.', '/', '-'], ['', '', ''], $numProtocoloFormatado);
+                            $retornoWsLinha = $this->executarServicoConsultaManifestacoes($this->urlWebServiceEOuv, $this->token, null, null, $numProtocoloSemFormatacao, $this->idRelatorioImportacao);
+                            $linkDetalheManifestacao = $retornoWsLinha[0]['Links'][0]['href'];
+                            $arrDetalheManifestacao = MdCguEouvWS::apiRestRequest($linkDetalheManifestacao, $this->token, 2);
+
+                            $debugLocal && LogSEI::getInstance()->gravar('Importando Recurso processo: ' . $numProtocoloFormatado . ' | tipo: ' . $tipo_recurso);
+
+
+                            /**
+                             * Verificar o tipo de recurso de for diferente de segunda instãncia, trazer todos os recursos para o documento pdf
+                             */
+                            if ($tipo_recurso <> 'R1') {
+                                $arrRecursosManifestacaoComAnteriores = $this->executarServicoConsultaRecursos($this->urlWebServiceESicRecursos, $this->token, null, null, $numProtocoloSemFormatacao);
+                                $this->gerarPDFDocumentoESic($arrDetalheManifestacao, $arrRecursosManifestacaoComAnteriores, $objProtocoloDTOExistente->getDblIdProtocolo(), $tipo_recurso);
+                            } else {
+                                $this->gerarPDFDocumentoESic($arrDetalheManifestacao, $arrRecursosManifestacao, $objProtocoloDTOExistente->getDblIdProtocolo(), $tipo_recurso);
+                            }
+
+                            $this->gravarLogLinha($numProtocoloFormatado, $this->idRelatorioImportacao, 'Recurso tipo ' . $tipo_recurso . ' com protocolo ' . $numProtocoloFormatado . ' importado com sucesso com ' . $anexoCount . ' anexos incluidos no protocolo.', 'S', $tipo_recurso, $dataPrazoAtendimento);
+                            $debugLocal && LogSEI::getInstance()->gravar('Importando Recurso processo: ' . $numProtocoloFormatado . ' | tipo: ' . $tipo_recurso . 'depois de gravar log ?!');
+                            // $this->gravarLogLinha($numProtocoloFormatado, $idRelatorioImportacao, 'Recurso com protocolo ' . $numProtocoloFormatado . ' importado com sucesso com ' . $anexoCount . ' anexos incluidos no protocolo.', 'S', $tipoManifestacao, $dataPrazoAtendimento);
+                            LogSEI::getInstance()->gravar('Módulo Integração FalaBR - Importação de Recurso ' . $numProtocoloFormatado . ': total de  Anexos configurados: ' . $anexoCount, InfraLog::$INFORMACAO);
+
+                            // Carregar anexos
+                            if (count($arrRecursosManifestacao['anexos']) > 0) {
+                                $this->gerarAnexosProtocolo($arrRecursosManifestacao['anexos'], $numProtocoloFormatado, $tipoManifestacao, $objProtocoloDTOExistente->getDblIdProtocolo());
+                            }
+
+                            try {
+                                $objEntradaEnviarProcesso = new EntradaEnviarProcessoAPI();
+                                $objEntradaEnviarProcesso->setIdProcedimento($objProtocoloDTOExistente->getDblIdProtocolo());
+                                $objEntradaEnviarProcesso->setUnidadesDestino([$unidadeDestino]);
+                                $objEntradaEnviarProcesso->setSinManterAbertoUnidade('S');
+                                $objEntradaEnviarProcesso->setSinEnviarEmailNotificacao('S');
+                                $objEntradaEnviarProcesso->setSinReabrir('S');
+
+                                $objSeiRN = new SeiRN();
+                                $objSeiRN->enviarProcesso($objEntradaEnviarProcesso);
+                                LogSEI::getInstance()->gravar('Módulo Integração FalaBR - (Recurso tipo ' . $tipo_recurso . ') Processo ' . $numProtocoloFormatado . ' enviado para unidade ' . $this->idUnidadeRecursoPrimeiraInstancia, InfraLog::$INFORMACAO);
+
+                            } catch (Exception $e) {
+                                LogSEI::getInstance()->gravar('Módulo Integração FalaBR - (Recurso tipo ' . $tipo_recurso . ') Não foi possivel abrir o Processo ' . $numProtocoloFormatado . ' na unidade ' . $this->idUnidadeRecursoPrimeiraInstancia . ' - erro: ' . $e, InfraLog::$INFORMACAO);
+                            }
+                        } else {
+                            $this->gravarLogLinha($numProtocoloFormatado, $this->idRelatorioImportacao, 'Sem recursos novos.', 'S', $ultimoTipoRecursoImportado, $dataPrazoAtendimento);
+                        }
+                    } catch (Exception $e) {
+                        $debugLocal && LogSEI::getInstance()->gravar('[executarImportacaoLinhaRecursos] Erro importando anexo do recruso');
+                        $this->gravarLogLinha($numProtocoloFormatado, $this->idRelatorioImportacao, 'Erro na gravação recurso: ' . $e, 'N', $tipoManifestacao);
+                    }
+                } else {
+                    $debugLocal && LogSEI::getInstance()->gravar('[executarImportacaoLinhaRecursos] Não importou recurso pois o prazo de atendimento é igual e não faz nada.. não atualiza o log para não atualizar a data do novo prazo nem o tipo de recurso');
+                    // Se não houve alteração na data 'PrazoAtendimento' retornar log
+                    $this->gravarLogLinha($numProtocoloFormatado, $this->idRelatorioImportacao, 'Já existe um recurso (e-Sic) do tipo "' . $this->verificaTipo($arrRecursosManifestacao, 'R') . '" para este protocolo e não há alteração para nova importação.', 'S', $ultimoTipoRecursoImportado, $objUltimaDataPrazoAtendimento->getDthDthPrazoAtendimento());
+                }
+            }
+        } else {
+            $this->gravarLogLinha($numProtocoloFormatado, $this->idRelatorioImportacao, 'Existe recurso para o processo ' . $numProtocoloFormatado . ', porém este processo não existe no SEI. Provavelmente é um processo antes da data de início de utilização do módulo ou o Tipo de Manifestação do FalaBR não foi registrada para este módulo.', 'S', $tipoManifestacao);
+        }
+    }
+
+    public function limparErrosParaNup($numProtocoloComErro){
+        $objEouvRelatorioImportacaoDetalheDTO = new MdCguEouvRelatorioImportacaoDetalheDTO();
+        $objEouvRelatorioImportacaoDetalheDTO->retTodos(true);
+        $objEouvRelatorioImportacaoDetalheDTO->setStrProtocoloFormatado($numProtocoloComErro);
+        $objEouvRelatorioImportacaoDetalheDTO->setStrSinSucesso('N');
+
+        $objEouvRelatorioImportacaoDetalheRN = new MdCguEouvRelatorioImportacaoDetalheRN();
+        $objListaErros = $objEouvRelatorioImportacaoDetalheRN->listar($objEouvRelatorioImportacaoDetalheDTO);
+        foreach($objListaErros as $erro){
+            $erro->setStrSinSucesso('C');
+            $objEouvRelatorioImportacaoDetalheRN->alterar($erro);
+        }
+
+    }
+
+    public function gerarPDFPedidoInicial($retornoWsLinha)
+    {
+        $mdCguEouvGerarPdfInicial = new MdCguEouvGerarPdfInicial($retornoWsLinha);
+        $pdf = $mdCguEouvGerarPdfInicial->gerarPdfInicial();
+
+        $objAnexoRN = new AnexoRN();
+        $strNomeArquivoInicialUpload = $objAnexoRN->gerarNomeArquivoTemporario();
+
+        $pdf->Output(DIR_SEI_TEMP . "/" . $strNomeArquivoInicialUpload . ".pdf", "F");
+
+        //Renomeia tirando a extensï¿½o para o SEI trabalhar o Arquivo
+        rename(DIR_SEI_TEMP . "/" . $strNomeArquivoInicialUpload . ".pdf", DIR_SEI_TEMP . "/" . $strNomeArquivoInicialUpload);
+
+        $objDocumentoManifestacao = new DocumentoAPI();
+        $objDocumentoManifestacao->setTipo('R');
+        $objDocumentoManifestacao->setIdSerie($this->idTipoDocumentoAnexoDadosManifestacao);
+        $objDocumentoManifestacao->setData($retornoWsLinha['DataCadastro']);
+        $objDocumentoManifestacao->setNomeArquivo('RelatórioDadosManifestação.pdf');
+        $objDocumentoManifestacao->setConteudo(base64_encode(file_get_contents(DIR_SEI_TEMP . "/" . $strNomeArquivoInicialUpload)));
+
+        return $objDocumentoManifestacao;
+    }
+
+    public function gerarPDFDocumentoESic($retornoWsLinha, $retornoWsRecursos = null, $IdProtocolo = false, $tipo_recurso = '')
+    {
+        /**
+         * Testa acessar dado da manifestação se não conseguir salva log
+         */
+        try {
+            $IdTipoManifestacao = $retornoWsLinha['TipoManifestacao']['IdTipoManifestacao'];
+        } catch (Exception $e) {
+            $this->gravarLogLinha($IdProtocolo ? $IdProtocolo : 'n/a', $idRelatorioImportacao, substr('ERRO-esic|' . $retornoWsLinha . '|' . $e, 0, 500), 'N');
+            return;
+        }
+
+        $pdf = MdCguEouvGerarPdfEsic::gerarPdf($retornoWsLinha, $retornoWsRecursos, $this->ocorreuErroAdicionarAnexo);
+        $objAnexoRN = new AnexoRN();
+        $strNomeArquivoInicialUpload = $objAnexoRN->gerarNomeArquivoTemporario();
+
+        $pdf->Output(DIR_SEI_TEMP . "/" . $strNomeArquivoInicialUpload . ".pdf", "F");
+
+        //Renomeuia tirando a extencaoo para o SEI trabalhar o Arquivo
+        rename(DIR_SEI_TEMP . "/" . $strNomeArquivoInicialUpload . ".pdf", DIR_SEI_TEMP . "/" . $strNomeArquivoInicialUpload);
+
+        $objDocumentoManifestacao = new DocumentoAPI();
+        $objDocumentoManifestacao->setTipo('R');
+        if ($IdProtocolo && $IdProtocolo <> '') {
+            $objDocumentoManifestacao->setIdProcedimento($IdProtocolo);
+        }
+        if ($tipo_recurso == 'R1') {
+            $nomeDocumentoArvore = 'Primeira Instância';
+        } elseif ($tipo_recurso == 'R2') {
+            $nomeDocumentoArvore = 'Segunda Instância';
+        } elseif ($tipo_recurso == 'R3' || $tipo_recurso == 'RC') {
+            $nomeDocumentoArvore = 'Terceira Instância';
+        } elseif ($tipo_recurso == 'PR') {
+            $nomeDocumentoArvore = 'Pedido Revisão';
+        } else {
+            $nomeDocumentoArvore = 'Pedido Inicial';
+        }
+
+        $objDocumentoManifestacao->setNumero($nomeDocumentoArvore);
+        $objDocumentoManifestacao->setIdSerie($this->idTipoDocumentoAnexoDadosManifestacao);
+        $objDocumentoManifestacao->setData($retornoWsLinha['DataCadastro']);
+        $objDocumentoManifestacao->setNomeArquivo('RelatorioDadosManifestacao.pdf');
+        $objDocumentoManifestacao->setConteudo(base64_encode(file_get_contents(DIR_SEI_TEMP . "/" . $strNomeArquivoInicialUpload)));
+
+        if ($IdProtocolo && $IdProtocolo <> '') {
+            $objSEIRN = new SeiRN();
+            $objSEIRN->incluirDocumento($objDocumentoManifestacao);
+        }
+
+        return $objDocumentoManifestacao;
+    }
+
+    public function gerarAnexosProtocolo($arrAnexosManifestacao, $numProtocoloFormatado, $tipoManifestacao = 'P', $IdProtocolo = false)
+    {
+        /**********************************************************************************************************************************************
+         * Início da importação de anexos de cada protocolo
+         * Desativado momentaneamente
+         */
+
+
+        $arrAnexosAdicionados = array();
+        $intTotAnexos = count($arrAnexosManifestacao);
+
+        if($intTotAnexos == 0){
+            //Não encontrou anexos..
+            return $arrAnexosAdicionados;
+        }
+
+        //Trata as extensÃµes permitidas
+        $objArquivoExtensaoDTO = new ArquivoExtensaoDTO();
+        $objArquivoExtensaoDTO->retNumIdArquivoExtensao();
+        $objArquivoExtensaoDTO->retStrExtensao();
+        $objArquivoExtensaoDTO->retStrDescricao();
+        $objArquivoExtensaoDTO->retNumTamanhoMaximo();
+        $objArquivoExtensaoRN = new ArquivoExtensaoRN();
+        $arrObjArquivoExtensaoDTO = $objArquivoExtensaoRN->listar($objArquivoExtensaoDTO);
+        $arrExtensoesPermitidas = array();
+
+        foreach($arrObjArquivoExtensaoDTO as $extensao){
+            array_push($arrExtensoesPermitidas, strtoupper ($extensao->getStrExtensao()));
+        }
+
+        foreach ($arrAnexosManifestacao as $retornoWsAnexoLista) {
+
+            foreach ($this->verificaRetornoWS($retornoWsAnexoLista) as $retornoWsAnexoLinha) {
+                try {
+
+                    $strNomeArquivoOriginal = $retornoWsAnexoLinha['NomeArquivo'];
+                    if ($strNomeArquivoOriginal == null) {
+                        $strNomeArquivoOriginal = $retornoWsAnexoLinha['nomeArquivo'];
+                    }
+
+                    // Ajustamos aqui o nome do arquivo limitado a 50 caracteres
+                    $strNomeArquivoOriginal = substr($strNomeArquivoOriginal, -50, 50);
+
+                    $ext = strtoupper(pathinfo($strNomeArquivoOriginal, PATHINFO_EXTENSION));
+                    $intIndexExtensao = array_search($ext, $arrExtensoesPermitidas);
+
+                    if (is_numeric($intIndexExtensao)) {
+                        $objAnexoRN = new AnexoRN();
+                        $strNomeArquivoUpload = $objAnexoRN->gerarNomeArquivoTemporario();
+
+                        $fp = fopen(DIR_SEI_TEMP . '/' . $strNomeArquivoUpload, 'w');
+
+                        //Busca o conteúdo do Anexo
+                        $arrDetalheAnexoManifestacao = MdCguEouvWS::apiRestRequest($retornoWsAnexoLinha['Links'][0]['href'], $this->token, 3);
+
+                        $strConteudoCodificado = $arrDetalheAnexoManifestacao['ConteudoZipadoEBase64'];
+
+                        $binConteudoDecodificado = '';
+                        for ($i = 0; $i < ceil(strlen($strConteudoCodificado) / 256); $i++) {
+                            $binConteudoDecodificado = $binConteudoDecodificado . base64_decode(substr($strConteudoCodificado, $i * 256, 256));
+                        }
+
+                        $binConteudoUnzip = $this->gzdecode($binConteudoDecodificado);
+
+                        fwrite($fp, $binConteudoUnzip);
+                        fclose($fp);
+
+                        $objAnexoManifestacao = new DocumentoAPI();
+
+                        if ($IdProtocolo && $IdProtocolo <> '') {
+                            $objAnexoManifestacao->setIdProcedimento($IdProtocolo);
+                        }
+                        $objAnexoManifestacao->setTipo('R');
+                        $objAnexoManifestacao->setIdSerie($this->idTipoDocumentoAnexoDadosManifestacao);
+                        $objAnexoManifestacao->setData(InfraData::getStrDataHoraAtual());
+                        $objAnexoManifestacao->setNomeArquivo($strNomeArquivoOriginal);
+                        $objAnexoManifestacao->setNumero($strNomeArquivoOriginal);
+                        $objAnexoManifestacao->setConteudo(base64_encode(file_get_contents(DIR_SEI_TEMP . '/' . $strNomeArquivoUpload)));
+
+                        if (!$this->hashDuplicado(DIR_SEI_TEMP . '/' . $strNomeArquivoUpload, $numProtocoloFormatado)) {
+                            if ($IdProtocolo && $IdProtocolo <> '') {
+                                $objSEIRN = new SeiRN();
+                                $objSEIRN->incluirDocumento($objAnexoManifestacao);
+                            }
+                            array_push($arrAnexosAdicionados, $objAnexoManifestacao);
+                        }
+                    } else {
+                        $ocorreuErroAdicionarAnexo = true;
+                        LogSEI::getInstance()->gravar('Importação de Manifestação ' . $numProtocoloFormatado . ': Arquivo ' . $strNomeArquivoOriginal . ' possui extensão inválida.', InfraLog::$INFORMACAO);
+                        continue;
+                    }
+                }
+                catch(Exception $e){
+                    $ocorreuErroAdicionarAnexo = true;
+                    $strMensagemErroAnexos = $strMensagemErroAnexos . " " . $e;
+                }
+            }
+
+            if($ocorreuErroAdicionarAnexo==true){
+                $this->gravarLogLinha($numProtocoloFormatado, $this->idRelatorioImportacao, 'Um ou mais documentos anexos não foram importados corretamente: ' . $strMensagemErroAnexos, 'S', $tipoManifestacao);
+            }
+        }
+
+        return $arrAnexosAdicionados;
+    }
+
+    public function excluirProcessoComErro($idProcedimento){
+
+        try{
+            $objProcedimentoExcluirDTO = new ProcedimentoDTO();
+            $objProcedimentoExcluirDTO->setDblIdProcedimento($idProcedimento);
+            $objProcedimentoRN = new ProcedimentoRN();
+            $objProcedimentoRN->excluirRN0280($objProcedimentoExcluirDTO);
+            ProcedimentoINT::removerProcedimentoVisitado($idProcedimento);
+        }catch(Exception $e){
+            PaginaSEI::getInstance()->processarExcecao($e);
+        }
+
+    }
+
+    public static function formatarProcesso($strProcesso) {
+
+        $strProcesso = InfraUtil::retirarFormatacao($strProcesso);
+
+        if (strlen($strProcesso)==0){
+            return '';
+        }
+
+        if (strlen($strProcesso) == 17){
+            $strProcesso = substr($strProcesso,0,5).".".
+                substr($strProcesso,5,6)."/".
+                substr($strProcesso,11,4)."-".
+                substr($strProcesso,15,2);
+        }
+        return $strProcesso;
+    }
+
+    /**
+     * Verifica se já existe um Protocolo no SEI com o número (NUP)
+     */
+    public function verificarProtocoloExistente($numProtocoloFormatado)
+    {
+        $objProtocoloDTOExistente = new ProtocoloDTO();
+        $objProtocoloRNExistente = new ProtocoloRN();
+        $objProtocoloDTOExistente->retDblIdProtocolo();
+        $objProtocoloDTOExistente->retStrProtocoloFormatado();
+        $objProtocoloDTOExistente->setStrProtocoloFormatado($this->formatarProcesso($numProtocoloFormatado));
+        $objProtocoloDTOExistente = $objProtocoloRNExistente->consultarRN0186($objProtocoloDTOExistente);
+
+        return $objProtocoloDTOExistente;
+    }
+
+    /**
+     * Verifica se já existe o hash do arquivo na tabela anexo coluna hash
+     *
+     * @param $strArquivo
+     * @return bool
+     * @throws InfraException
+     */
+    public function hashDuplicado($strArquivo, $numProtocoloFormatado)
+    {
+        // Verifica hash do arquivo
+        $hash = md5_file($strArquivo);
+
+        // Select na tabela Anexe com o hash Criado
+        $consulta = new MdCguEouvConsultarHashBD($this->getObjInfraIBanco());
+        $res = $consulta->consultarHash($hash, $numProtocoloFormatado);
+
+        return count($res) > 0;
+    }
+
+    /**
+     * Função para simular login
+     *
+     * @param $siglaSistema
+     * @param $idServico
+     * @param $idUnidade
+     */
+    public static function simulaLogin($siglaSistema, $idServico, $idUnidade)
+    {
         try {
 
-            //Retorna dados da Última execução com Sucesso
-            $objUltimaExecucao = MdCguEouvAgendamentoINT::retornarUltimaExecucaoSucesso();
+            InfraDebug::getInstance()->gravar(__METHOD__);
+            InfraDebug::getInstance()->gravar('SIGLA SISTEMA:'.$siglaSistema);
+            InfraDebug::getInstance()->gravar('IDENTIFICACAO SERVICO:'.$idServico);
+            InfraDebug::getInstance()->gravar('ID UNIDADE:'.$idUnidade);
 
-            if ($objUltimaExecucao != null) {
-                $ultimaDataExecucao = $objUltimaExecucao->getDthDthPeriodoFinal();
-            } //Primeira execução ou nenhuma executada com sucesso
-            else {
-                $ultimaDataExecucao = $this->dataInicialImportacaoManifestacoes;
-            }
+            SessaoSEI::getInstance(false);
 
-            $semManifestacoesEncontradas = true;
-            $qtdManifestacoesNovas = 0;
-            $qtdManifestacoesAntigas = 0;
-            $objEouvRelatorioImportacaoDTO = $this->gravarLogImportacao($ultimaDataExecucao, $dataAtual);
-            $idRelatorioImportacao = $objEouvRelatorioImportacaoDTO->getNumIdRelatorioImportacao();
-            $objEouvRelatorioImportacaoRN = new MdCguEouvRelatorioImportacaoRN();
-            $SinSucessoExecucao = 'N';
-            $textoMensagemErroToken = '';
+            $objServicoDTO = self::obterServico($siglaSistema, $idServico);
 
-            $retornoWs = $mdCguEouvWS->executarServicoConsultaManifestacoes($this->urlWebServiceEOuv, $this->token, $ultimaDataExecucao, $dataAtual, null, $idRelatorioImportacao);
-
-            //Caso retornado algum erro
-            if (is_string($retornoWs)) {
-                if (strpos($retornoWs, 'Invalidado') !== false) {
-                    //Tenta gerar novo token
-                    $tokenValido = MdCguEouvWS::apiValidarToken($this->urlWebServiceEOuv, $this->usuarioWebService, $this->senhaUsuarioWebService, $this->client_id, $this->client_secret);
-
-                    if (isset($tokenValido['error'])) {
-                        $textoMensagemErroToken = 'Não foi possível validar o Token de acesso aos WebServices do E-ouv. <br>Verifique as informações de Usuário, Senha, Client_Id e Client_Secret nas configurações de Parâmetros do Módulo';
-
-                    } elseif (isset($tokenValido['access_token'])) {
-                        $mdCguEouvWS->gravarParametroToken($tokenValido['access_token']);
-                        $token = $tokenValido['access_token'];
-
-                        //Chama novamente a execução da ConsultaManifestacao que deu errado por causa do Token
-                        $retornoWs = $mdCguEouvWS->executarServicoConsultaManifestacoes($this->urlWebServiceEOuv, $token, $ultimaDataExecucao, $dataAtual, null, $idRelatorioImportacao);
-                    }
-                }
-            }
-
-            if ($textoMensagemErroToken == '') {
-                $arrComErro = $mdCguEouvWS->obterManifestacoesComErro($this->urlWebServiceEOuv, $token, $ultimaDataExecucao, $dataAtual, $idRelatorioImportacao);
-
-                $arrManifestacoes = array();
-
-                if (is_array($retornoWs)) {
-                    // Filtra as manifestações e-Ouv
-                    $arrManifestacoes = array_filter($retornoWs, function($manifestacao) {
-                        return $manifestacao['TipoManifestacao']['IdTipoManifestacao'] <> 8;
-                    });
-                    $qtdManifestacoesNovas = count($arrManifestacoes);
-                }
-
-                if (is_array($arrComErro)) {
-                    $qtdManifestacoesAntigas = count($arrComErro);
-                    $arrManifestacoes = array_merge($arrManifestacoes, $arrComErro);
-                }
-
-                if (count($arrManifestacoes) > 0) {
-                    $semManifestacoesEncontradas = false;
-                    foreach ($arrManifestacoes as $retornoWsLinha) {
-                        $mdCguEouvWS->executarImportacaoLinha($retornoWsLinha, 'P', $idRelatorioImportacao);
-                    }
-                }
-
-                $textoMensagemFinal = 'Execução Finalizada com Sucesso!';
-                $SinSucessoExecucao = 'S';
-
-                if ($semManifestacoesEncontradas) {
-                    $textoMensagemFinal = $textoMensagemFinal . ' Não foram encontradas manifestações para o período.';
-                } else {
-                    $textoMensagemFinal = $textoMensagemFinal . '<br>Quantidade de Manifestações novas encontradas (e-Ouv): ' . $qtdManifestacoesNovas . '<br>Quantidade de Manifestações encontadas que ocorreram erro em outras importações: ' . $qtdManifestacoesAntigas;
-                }
-
-                if ($this->ocorreuErroEmProtocolo) {
-                    $textoMensagemFinal = $textoMensagemFinal . '<br> Ocorreram erros em 1 ou mais protocolos.';
-                }
+            if ($idUnidade!=null) {
+                $objUnidadeDTO = self::obterUnidade($idUnidade,null);
             } else {
-                $textoMensagemFinal = $textoMensagemErroToken;
+                $objUnidadeDTO = null;
             }
-
-            //Grava a execução com sucesso se tiver corrido tudo bem
-            $objEouvRelatorioImportacaoDTO2 = new MdCguEouvRelatorioImportacaoDTO();
-
-            $objEouvRelatorioImportacaoDTO2->setNumIdRelatorioImportacao($objEouvRelatorioImportacaoDTO->getNumIdRelatorioImportacao());
-            $objEouvRelatorioImportacaoDTO2->setStrSinSucesso($SinSucessoExecucao);
-            $objEouvRelatorioImportacaoDTO2->setStrDeLogProcessamento($textoMensagemFinal);
-            $objEouvRelatorioImportacaoRN->alterar($objEouvRelatorioImportacaoDTO2);
-
-
+            SessaoSEI::getInstance()->simularLogin(null, null, $objServicoDTO->getNumIdUsuario(), $objUnidadeDTO->getNumIdUnidade());
 
         } catch(Exception $e) {
-
-            $objEouvRelatorioImportacaoDTO3 = new MdCguEouvRelatorioImportacaoDTO();
-            $objEouvRelatorioImportacaoDTO3->setNumIdRelatorioImportacao($objEouvRelatorioImportacaoDTO->getNumIdRelatorioImportacao());
-            $strMensagem = 'Ocorreu um erro no processamento:' . $e;
-            $strMensagem = substr($strMensagem, 0, 500);
-            $objEouvRelatorioImportacaoDTO3->setStrDeLogProcessamento($strMensagem);
-
-            $objEouvRelatorioImportacaoRN->alterar($objEouvRelatorioImportacaoDTO3);
-
+            LogSEI::getInstance()->gravar('Ocorreu erro simular Login.'.$e);
             PaginaSEI::getInstance()->processarExcecao($e);
-        } finally {
-            //Restaura a sessão
-            SessaoSEI::getInstance()->setBolHabilitada($isBolHabilitada);
         }
     }
 
     /**
-     * Função para importar as manifestações e-Sic do FalaBR (tipo 8)
+     * Verifica o tipo de Recuso com base na API do FalaBR
+     *
+     * - IdInstanciaRecurso
+     * - 1 = primeira instância
+     * - 2 = segunda instância
+     *
+     * @param null $recursos
+     * @return string
+     *
+     * - 'P' - Padrão, não possui recursos de primeira ou segunda instância
+     * - 'R1' - Recurso de primeira instância
+     * - 'R2' - Recurso de segunda instância
      */
-    public function executarImportacaoManifestacaoESic()
+    public function verificaTipo($recursos = null, $default_response = 'P')
+    {
+        $response = $default_response;
+        if (isset($recursos)) {
+            if (isset($recursos['instancia'])) {
+                $response = $this->checkTipoRecurso($recursos);
+            } else {
+                foreach ($recursos as $recurso) {
+                    if ($this->checkTipoRecurso($recurso)) {
+                        $response = $this->checkTipoRecurso($recurso);
+                        break;
+                    }
+                }
+
+            }
+        }
+
+        return $response;
+    }
+
+    public function checkTipoRecurso($recurso)
+    {
+        if ($recurso['instancia']['IdInstanciaRecurso'] == 6) {
+            return 'PR'; // Pedido Revisão
+        }
+        if ($recurso['instancia']['IdInstanciaRecurso'] == 7) {
+            return 'R3'; // Recurso 3 instância
+        }
+        if ($recurso['instancia']['IdInstanciaRecurso'] == 4) {
+            return 'RE'; // Reclamação
+        }
+        if ($recurso['instancia']['IdInstanciaRecurso'] == 3) {
+            return 'RC'; // Recurso CGU
+        }
+        if ($recurso['instancia']['IdInstanciaRecurso'] == 2) {
+            return 'R2'; // Recurso 2 instância
+        }
+        if ($recurso['instancia']['IdInstanciaRecurso'] == 1) {
+            return 'R1'; // Recurso 1 instância
+        }
+
+        return 'R';
+    }
+
+    /**
+     * Verifica se existe recurso 'posterior' cadastrado
+     *
+     * - Posterior está entre aspas pq o recurso deve seguir uma órdem cronológica para se adequar à importação dos
+     * dados no SEI
+     *
+     * @param $idRelatorioImportacao
+     * @param $numProtocolo
+     * @param $tipoManifestacao
+     * @return bool|void
+     */
+    public function permiteImportacaoRecursoAtual($tipoManifestacaoAtual, $ultimoTipoRecursoImportado)
     {
         $debugLocal = false;
 
-        // Log
-        LogSEI::getInstance()->gravar('Rotina de Importação de Manifestações do FalaBR (e-Sic)', InfraLog::$INFORMACAO);
+        $debugLocal && LogSEI::getInstance()->gravar('[permiteImportacaoRecursoAtual] Verificando se existe algum recurso anterior');
 
-        // Lista parâmetros
-        $objEouvParametroDTO = new MdCguEouvParametroDTO();
-        $objEouvParametroDTO->retTodos();
+        // Se ja existir no log um recurso anterior verifica se o novo recurso e 'superior' ao já registrado
+        if ($tipoManifestacaoAtual) {
 
-        // Busca parâmetros do banco de dados da tabela md_eouv_parametros
-        $objEouvParametroRN = new MdCguEouvParametroRN();
-        $arrObjEouvParametroDTO = $objEouvParametroRN->listarParametro($objEouvParametroDTO);
-        $numRegistros = count($arrObjEouvParametroDTO);
+            $debugLocal && LogSEI::getInstance()->gravar('[permiteImportacaoRecursoAtual] Existe log, validando o tipo de manifestação: ' . $tipoManifestacaoAtual . ' para o anteior existente: ' . $ultimoTipoRecursoImportado);
 
-        $this->preencheVariaveis($numRegistros, $arrObjEouvParametroDTO);
+            /**
+             * [CUIDADO] Nâo é possível utilizar o 'switch > case' aqui - não sei o por quê, mas não funciona....  @study (??)
+             */
+
+            /**
+             * Para criar um R1 (Recurso de Primeira Instância) pode existir somente PR (Pedido de Revisão),
+             * R (Pedido Inicial do e-Sic)
+             */
+            if ($tipoManifestacaoAtual == 'R1' && in_array($ultimoTipoRecursoImportado, ['R2', 'RC', 'R3'])) {
+                $debugLocal && LogSEI::getInstance()->gravar('[permiteImportacaoRecursoAtual] Deve bloquear a criação deste recurso! tipoAtual: ' . $tipoManifestacaoAtual . ' - tipoAnterior: ' . $ultimoTipoRecursoImportado);
+                return 'bloquear';
+            }
+
+            /**
+             * Para criar um R2 (Recurso de Segunda Instância) pode existir somente R1 (Recurso de Primeira Instância),
+             * PR (Pedido de Revisão), R (Pedido Inicial do e-Sic)
+             */
+            if ($tipoManifestacaoAtual == 'R2' && in_array($ultimoTipoRecursoImportado, ['RC', 'R3'])) {
+                $debugLocal && LogSEI::getInstance()->gravar('[permiteImportacaoRecursoAtual] Deve bloquear a criação deste recurso! tipoAtual: ' . $tipoManifestacaoAtual . ' - tipoAnterior: ' . $ultimoTipoRecursoImportado);
+                return 'bloquear';
+            }
+
+            /**
+             * Se for tipo 4 - Reclamação - não importar
+             */
+            if ($tipoManifestacaoAtual == 'RE') {
+                $debugLocal && LogSEI::getInstance()->gravar('[permiteImportacaoRecursoAtual] Deve bloquear a criação deste recurso! tipoAtual: ' . $tipoManifestacaoAtual . ' - tipoAnterior: ' . $ultimoTipoRecursoImportado);
+                return 'bloquear';
+            }
+
+            /**
+             * Para criar um PR (Pedido de Revisão) pode existir somente R (Pedido Inicial do e-Sic)
+             */
+            if ($tipoManifestacaoAtual == 'PR' && in_array($ultimoTipoRecursoImportado, ['R1', 'R2', 'RC', 'R3'])) {
+                $debugLocal && LogSEI::getInstance()->gravar('[permiteImportacaoRecursoAtual] Deve bloquear a criação deste recurso! tipoAtual: ' . $tipoManifestacaoAtual . ' - tipoAnterior: ' . $ultimoTipoRecursoImportado);
+                return 'bloquear';
+            }
+        }
 
         /**
-         * Função para buscar o 'restante' do token sem o limite de 255 caracteres do SEI
+         * Se existir algo na tabela, porém, não estiver definido na regra acima ou se não existir nenhum registro na
+         * tabela, a importação será permitida
+         * [CUIDADO] Caso haja duplicidade na importação, pode haver algum tipo de recurso não mapeado no campo
+         * "instancia": { "IdInstanciaRecurso": ## > na API do FalaBR
          */
-        $tokenPart2 = BancoSEI::getInstance()->consultarSql("select substring(de_valor_parametro, 256, 455) from md_eouv_parametros where no_parametro='TOKEN';")[0]['computed'];
-        $this->token = $this->token . $tokenPart2;
-
-        // Busca parâmetros do banco de dados da tabela infra_parametros
-        $dataAtual = InfraData::getStrDataHoraAtual();
-
-        $isBolHabilitada = SessaoSEI::getInstance(false)->isBolHabilitada();
-        SessaoSEI::getInstance()->setBolHabilitada(false);
-
-        $mdCguEouvWS = new MdCguEouvWS($this->urlWebServiceEOuv, $this->urlWebServiceESicRecursos, $this->idTipoDocumentoAnexoDadosManifestacao,
-            $this->idUnidadeOuvidoria, $this->idUnidadeEsicPrincipal, $this->idUnidadeRecursoPrimeiraInstancia, $this->idUnidadeRecursoSegundaInstancia,
-            $this->idUnidadeRecursoTerceiraInstancia, $this->idUnidadeRecursoPedidoRevisao, $this->token);
-        // Simula login inicial
-        $mdCguEouvWS->simulaLogin($this->siglaSistema, $this->identificacaoServico, $this->idUnidadeEsicPrincipal);
-
-        // Executa a importação dos dados
-        try {
-            //Retorna dados da Última execução com Sucesso
-            $objUltimaExecucao = MdCguEouvAgendamentoINT::retornarUltimaExecucaoSucesso('R');
-
-            if ($objUltimaExecucao != null) {
-                // Debug Logs
-                $debugLocal && LogSEI::getInstance()->gravar('$objUltimaExecuxao (e-Sic):' . $objUltimaExecucao->getDthDthPeriodoFinal());
-
-                $ultimaDataExecucao = $objUltimaExecucao->getDthDthPeriodoFinal();
-            } else {
-                // Debug Logs
-                $debugLocal && LogSEI::getInstance()->gravar('$objUltimaExecuxao (e-Sic) é NULL');
-
-                //Primeira execução ou nenhuma executada com sucesso
-                $ultimaDataExecucao = $this->dataInicialImportacaoManifestacoes;
-            }
-
-            $semManifestacoesEncontradas = true;
-            $qtdManifestacoesNovas = 0;
-            $qtdManifestacoesAntigas = 0;
-            $semRecursosEncontrados = true;
-            $qtdRecursosNovos = 0;
-
-            /**
-             * A função abaixo gravarLogImportacao recebe o tipo de manifestação 'R' (Recursos) para as manifestações do e-Sic
-             */
-            $objEouvRelatorioImportacaoDTO = $this->gravarLogImportacao($ultimaDataExecucao, $dataAtual, 'R');
-            $idRelatorioImportacao = $objEouvRelatorioImportacaoDTO->getNumIdRelatorioImportacao();
-            $objEouvRelatorioImportacaoRN = new MdCguEouvRelatorioImportacaoRN();
-            $SinSucessoExecucao = 'N';
-            $textoMensagemErroToken = '';
-
-            /**
-             * As funções abaixo fazem a busca no webservice dos dados a serem trabalhados na rotina de importação
-             */
-            $debugLocal && LogSEI::getInstance()->gravar('Iniciando a consulta inicial');
-
-            $retornoWs = $mdCguEouvWS->executarServicoConsultaManifestacoes($this->urlWebServiceEOuv, $this->token, $ultimaDataExecucao, $dataAtual, null, $idRelatorioImportacao);
-            $retornoWsRecursos = $mdCguEouvWS->executarServicoConsultaRecursos($this->urlWebServiceESicRecursos, $this->token, $ultimaDataExecucao, $dataAtual, null, $idRelatorioImportacao);
-
-            //Caso retornado algum erro - Manifestações e-Sic
-            if (is_string($retornoWs)) {
-                $debugLocal && LogSEI::getInstance()->gravar('Retorno da consulta $retornoWs é uma string: ' . $retornoWs);
-
-                if (strpos($retornoWs, 'Invalidado') !== false) {
-                    //Tenta gerar novo token
-                    $tokenValido = MdCguEouvWS::apiValidarToken($this->urlWebServiceEOuv, $this->usuarioWebService, $this->senhaUsuarioWebService, $this->client_id, $this->client_secret);
-
-                    if (isset($tokenValido['error'])) {
-                        $textoMensagemErroToken = 'Não foi possível validar o Token de acesso aos WebServices do E-ouv. <br>Verifique as informações de Usuário, Senha, Client_Id e Client_Secret nas configurações de Parâmetros do Módulo';
-
-                    } elseif (isset($tokenValido['access_token'])) {
-                        $this->gravarParametroToken($tokenValido['access_token']);
-                        $token = $tokenValido['access_token'];
-
-                        //Chama novamente a execução da ConsultaManifestacao que deu errado por causa do Token
-                        $retornoWs = $mdCguEouvWS->executarServicoConsultaManifestacoes($this->urlWebServiceEOuv, $token, $ultimaDataExecucao, $dataAtual, null, $idRelatorioImportacao);
-                        $retornoWsRecursos = $mdCguEouvWS->executarServicoConsultaRecursos($this->urlWebServiceESicRecursos, $token, $ultimaDataExecucao, $dataAtual, null, $idRelatorioImportacao);
-                    }
-                }
-            }
-
-            /**
-             * @todo - criar rotina para buscar recursos das manifestções com erro caso exista alguma na tabela de log
-             */
-
-            if ($textoMensagemErroToken == '') {
-
-                /**
-                 * @debug - manifestacao com erro
-                 * Comentar a linha abaixo para debugar um retorno manual
-                 */
-                $debugLocal && LogSEI::getInstance()->gravar('Inicia busca manifestação com erros');
-                $arrComErro = $mdCguEouvWS->obterManifestacoesComErro($this->urlWebServiceEOuv, $token, $ultimaDataExecucao, $dataAtual, $idRelatorioImportacao, 'R');
-
-                $arrManifestacoes = array();
-
-                if (is_array($retornoWs)) {
-                    // Filtra as manifestações e-Sic
-                    $arrManifestacoes = array_filter($retornoWs, function($manifestacao) {
-                        return $manifestacao['TipoManifestacao']['IdTipoManifestacao'] == 8;
-                    });
-                    $qtdManifestacoesNovas = count($arrManifestacoes);
-                    $debugLocal && LogSEI::getInstance()->gravar('Possui novas manifestações qtd: ' . $qtdManifestacoesNovas);
-                }
-
-                $arrRecursos = array();
-                if (isset($retornoWsRecursos) && is_array($retornoWsRecursos)) {
-                    $debugLocal && LogSEI::getInstance()->gravar('Possui recursos qtd: ' . count($retornoWsRecursos['Recursos']));
-                    $arrRecursos = $retornoWsRecursos['Recursos'];
-                    $qtdRecursosNovos = count($arrRecursos);
-                }
-
-                if (is_array($arrComErro)) {
-                    $debugLocal && LogSEI::getInstance()->gravar('Possui manifestações com erros - qtd: ' . count($arrComErro));
-                    $qtdManifestacoesAntigas = count($arrComErro);
-                    $arrManifestacoes = array_merge($arrManifestacoes, $arrComErro);
-                }
-
-                // Importa manifestações e-Sic
-                if (count($arrManifestacoes) > 0) {
-                    $semManifestacoesEncontradas = false;
-                    foreach ($arrManifestacoes as $retornoWsLinha) {
-                        $debugLocal && LogSEI::getInstance()->gravar('Inicia importação por Linha');
-                        $mdCguEouvWS->executarImportacaoLinha($retornoWsLinha, 'R',$idRelatorioImportacao);
-                    }
-                }
-
-                // Importa recursos e-Sic
-                if (count($arrRecursos) > 0) {
-                    $semRecursosEncontrados = false;
-                    foreach ($arrRecursos as $retornoWsLinha) {
-                        $debugLocal && LogSEI::getInstance()->gravar('Inicia importação por linha de Recursos - protocolo: ' . $retornoWsLinha['numProtocolo']);
-                        $mdCguEouvWS->executarImportacaoLinhaRecursos($retornoWsLinha);
-                    }
-                }
-
-                $textoMensagemFinal = 'Execução Finalizada com Sucesso!';
-                $SinSucessoExecucao = 'S';
-
-                if ($semManifestacoesEncontradas) {
-                    $textoMensagemFinal = $textoMensagemFinal . ' Não foram encontradas manifestações para o período.';
-                } else {
-                    $textoMensagemFinal = $textoMensagemFinal . '<br>Quantidade de Manifestações novas encontradas (e-Sic): ' . $qtdManifestacoesNovas . '<br>Quantidade de Manifestações encontadas que ocorreram erro em outras importações: ' . $qtdManifestacoesAntigas;
-                }
-
-                if ($semRecursosEncontrados) {
-                    $textoMensagemFinal = $textoMensagemFinal . ' Não foram encontrados recursos para o período.';
-                } else {
-                    $textoMensagemFinal = $textoMensagemFinal . '<br>Quantidade de Recursos novos encontrados (e-Sic): ' . $qtdRecursosNovos;
-                }
-
-                if ($this->ocorreuErroEmProtocolo) {
-                    $textoMensagemFinal = $textoMensagemFinal . '<br> Ocorreram erros em 1 ou mais protocolos.';
-                }
-            } else {
-                $textoMensagemFinal = $textoMensagemErroToken;
-            }
-
-            //Grava a execução com sucesso se tiver corrido tudo bem
-            $objEouvRelatorioImportacaoDTO2 = new MdCguEouvRelatorioImportacaoDTO();
-
-            $objEouvRelatorioImportacaoDTO2->setNumIdRelatorioImportacao($objEouvRelatorioImportacaoDTO->getNumIdRelatorioImportacao());
-            $objEouvRelatorioImportacaoDTO2->setStrSinSucesso($SinSucessoExecucao);
-            $objEouvRelatorioImportacaoDTO2->setStrDeLogProcessamento($textoMensagemFinal);
-            $objEouvRelatorioImportacaoDTO2->setStrTipManifestacao('R');
-            $objEouvRelatorioImportacaoRN->alterar($objEouvRelatorioImportacaoDTO2);
-
-            LogSEI::getInstance()->gravar('Finalizado a importção dos processos e-Sic - FalaBR');
-
-        } catch(Exception $e) {
-
-            $objEouvRelatorioImportacaoDTO3 = new MdCguEouvRelatorioImportacaoDTO();
-            $objEouvRelatorioImportacaoDTO3->setNumIdRelatorioImportacao($objEouvRelatorioImportacaoDTO->getNumIdRelatorioImportacao());
-            $strMensagem = 'Ocorreu um erro no processamento:' . $e;
-            $strMensagem = substr($strMensagem, 0, 500);
-            $objEouvRelatorioImportacaoDTO3->setStrDeLogProcessamento($strMensagem);
-            $objEouvRelatorioImportacaoRN->alterar($objEouvRelatorioImportacaoDTO3);
-
-            PaginaSEI::getInstance()->processarExcecao($e);
-
-        } finally {
-            //Restaura a Sessão
-            SessaoSEI::getInstance()->setBolHabilitada($isBolHabilitada);
-        }
+        $debugLocal && LogSEI::getInstance()->gravar('[permiteImportacaoRecursoAtual] Vai permitir a criação desse recurso!');
+        return 'permitir';
     }
 }
 ?>
